refactor: split SportsEventCenter into focused sub-components
