+++ src/app/[locale]/(platform)/sports/_components/SportsEventCenter.tsx
@@ -1,3533 +1,87 @@
 'use client'
 
-import type { SportsGamesMarketType, SportsLinePickerOption } from '@/app/[locale]/(platform)/sports/_components/SportsGamesCenter'
-import type { SportsRedeemModalGroup, SportsRedeemModalSection } from '@/app/[locale]/(platform)/sports/_components/SportsRedeemModal'
+import type {
+  EventSectionKey,
+  SportsEventCenterProps,
+} from '@/app/[locale]/(platform)/sports/_components/sports-event-center-types'
 import type {
   SportsGamesButton,
-  SportsGamesCard,
-  SportsGamesCardMarketView,
 } from '@/app/[locale]/(platform)/sports/_utils/sports-games-data'
-import type { OddsFormat } from '@/lib/odds-format'
-import type { SportsEventMarketViewKey } from '@/lib/sports-event-slugs'
-import type { SportsVertical } from '@/lib/sports-vertical'
-import type { UserPosition } from '@/types'
-import { useQuery } from '@tanstack/react-query'
-import { CheckIcon, ChevronLeftIcon, ChevronRightIcon, ShareIcon } from 'lucide-react'
+import { ChevronLeftIcon } from 'lucide-react'
 import { useLocale } from 'next-intl'
 import Image from 'next/image'
-import { useSearchParams } from 'next/navigation'
-import { Suspense, useCallback, useEffect, useMemo, useRef, useState, useSyncExternalStore } from 'react'
+import { Suspense, useMemo } from 'react'
 import EventBookmark from '@/app/[locale]/(platform)/event/[slug]/_components/EventBookmark'
-import { useOrderBookSummaries } from '@/app/[locale]/(platform)/event/[slug]/_components/EventOrderBook'
 import EventOrderPanelForm from '@/app/[locale]/(platform)/event/[slug]/_components/EventOrderPanelForm'
 import EventOrderPanelMobile from '@/app/[locale]/(platform)/event/[slug]/_components/EventOrderPanelMobile'
 import EventOrderPanelTermsDisclaimer
   from '@/app/[locale]/(platform)/event/[slug]/_components/EventOrderPanelTermsDisclaimer'
 import EventTabs from '@/app/[locale]/(platform)/event/[slug]/_components/EventTabs'
+import {
+  useActiveCardPriceMap,
+  useActiveMarketView,
+  useActiveTradeContext,
+  useAuxiliaryMarketCards,
+  useClaimGroupsBySection,
+  useDerivedActiveCard,
+  useEsportsSegmentTabKeySync,
+  useEsportsSegmentTabState,
+  useFormatButtonOdds,
+  useGroupedButtons,
+  useMarketSlugToButtonKey,
+  useOddsFormat,
+  useOrderStateSync,
+  useQuerySelection,
+  useRedeemModalState,
+  useSectionActions,
+  useSectionDerivedData,
+  useSelectedSectionButtons,
+  useSelectionState,
+  useSeriesSegmentPickerData,
+  useSeriesSpreadPickerSync,
+  useUserPositionsQuery,
+} from '@/app/[locale]/(platform)/sports/_components/sports-event-center-hooks'
+import {
+  headerIconButtonClass,
+} from '@/app/[locale]/(platform)/sports/_components/sports-event-center-types'
+import {
+  normalizeLivestreamUrl,
+  parseSportsScore,
+  resolveMoneylineButtonGridClass,
+  resolveTeamShortLabel,
+  shouldUseFullScoreboardHeroLabels,
+  sortSectionButtons,
+} from '@/app/[locale]/(platform)/sports/_components/sports-event-center-utils'
 import SportsEventAboutPanel from '@/app/[locale]/(platform)/sports/_components/SportsEventAboutPanel'
-import { buildLinePickerOptions, groupButtonsByMarketType, resolveButtonDepthStyle, resolveButtonOverlayStyle, resolveButtonStyle, resolveDefaultConditionId, resolveOrderPanelOutcomeAccentOverrides, resolveOrderPanelOutcomeLabelOverrides, resolvePreferredLinePickerButton, resolveSelectedButton, resolveSelectedMarket, resolveSelectedOutcome, resolveStableSpreadPrimaryOutcomeIndex, SportsGameDetailsPanel, SportsGameGraph, SportsOrderPanelMarketInfo } from '@/app/[locale]/(platform)/sports/_components/SportsGamesCenter'
+import SportsEventLiveStatusIcon from '@/app/[locale]/(platform)/sports/_components/SportsEventLiveStatusIcon'
+import SportsEventQuerySync from '@/app/[locale]/(platform)/sports/_components/SportsEventQuerySync'
+import SportsEventRelatedGames from '@/app/[locale]/(platform)/sports/_components/SportsEventRelatedGames'
+import SportsEventShareButton from '@/app/[locale]/(platform)/sports/_components/SportsEventShareButton'
+import {
+  resolveButtonDepthStyle,
+  resolveButtonOverlayStyle,
+  resolveButtonStyle,
+  resolveSelectedButton,
+  SportsGameDetailsPanel,
+  SportsGameGraph,
+  SportsOrderPanelMarketInfo,
+} from '@/app/[locale]/(platform)/sports/_components/SportsGamesCenter'
 import SportsLivestreamFloatingPlayer
   from '@/app/[locale]/(platform)/sports/_components/SportsLivestreamFloatingPlayer'
 import SportsRedeemModal from '@/app/[locale]/(platform)/sports/_components/SportsRedeemModal'
-import { buildMarketSlugSelectionSignature } from '@/app/[locale]/(platform)/sports/_utils/sports-event-selection'
-import {
-  resolveSportsAuxiliaryMarketGroupKey,
-  resolveSportsAuxiliaryMarketTitle,
-} from '@/app/[locale]/(platform)/sports/_utils/sports-games-data'
+import SportsSegmentNumberPicker from '@/app/[locale]/(platform)/sports/_components/SportsSegmentNumberPicker'
 import AppLink from '@/components/AppLink'
 import SiteLogoIcon from '@/components/SiteLogoIcon'
-import { Button } from '@/components/ui/button'
 import { useCurrentTimestamp } from '@/hooks/useCurrentTimestamp'
 import { useIsMobile } from '@/hooks/useIsMobile'
 import { useSiteIdentity } from '@/hooks/useSiteIdentity'
-import { ensureReadableTextColorOnDark } from '@/lib/color-contrast'
-import { ORDER_SIDE, OUTCOME_INDEX } from '@/lib/constants'
-import { fetchUserPositionsForMarket } from '@/lib/data-api/user'
 import { formatVolume } from '@/lib/formatters'
-import { resolveOutcomePriceCents, resolveOutcomeSelectionPriceCents } from '@/lib/market-pricing'
-import { formatOddsFromCents, ODDS_FORMAT_OPTIONS } from '@/lib/odds-format'
 import { shouldUseCroppedSportsTeamLogo } from '@/lib/sports-team-logo'
 import { getSportsVerticalConfig } from '@/lib/sports-vertical'
 import { cn } from '@/lib/utils'
 import { useOrder } from '@/stores/useOrder'
 import { useSportsLivestream } from '@/stores/useSportsLivestream'
 import { useUser } from '@/stores/useUser'
-
-type DetailsTab = 'orderBook' | 'graph' | 'about'
-type EventSectionKey = Extract<SportsGamesMarketType, 'moneyline' | 'spread' | 'total' | 'btts'>
-type EsportsLayoutTabKey = 'series' | `segment-${number}`
-
-interface SportsEventCenterProps {
-  card: SportsGamesCard
-  marketViewCards?: SportsGamesCardMarketView[]
-  relatedCards?: SportsGamesCard[]
-  marketContextEnabled?: boolean
-  sportSlug: string
-  sportLabel: string
-  initialMarketSlug?: string | null
-  initialMarketViewKey?: SportsEventMarketViewKey | null
-  vertical?: SportsVertical
-}
-
-interface SportsEventQuerySelection {
-  conditionId: string | null
-  outcomeIndex: number | null
-}
-
-interface AuxiliaryMarketPanel {
-  key: string
-  title: string
-  markets: SportsGamesCard['detailMarkets']
-  buttons: SportsGamesButton[]
-  volume: number
-  mapNumber: number | null
-}
-
-interface SportsSegmentNumberPickerOption {
-  key: string
-  label: string
-  number: number
-}
-
-const SECTION_ORDER: Array<{ key: EventSectionKey, label: string }> = [
-  { key: 'moneyline', label: 'Moneyline' },
-  { key: 'spread', label: 'Spread' },
-  { key: 'total', label: 'Totals' },
-  { key: 'btts', label: 'Both Teams to Score?' },
-]
-
-const headerIconButtonClass = `
-  size-10 rounded-sm border border-transparent bg-transparent text-foreground transition-colors
-  hover:bg-muted/80 focus-visible:ring-1 focus-visible:ring-ring md:h-9 md:w-9
-`
-const SPORTS_EVENT_ODDS_FORMAT_STORAGE_KEY = 'sports:event:odds-format'
-const EMPTY_QUERY_SELECTION: SportsEventQuerySelection = {
-  conditionId: null,
-  outcomeIndex: null,
-}
-
-function resolveInitialOddsFormat(): OddsFormat {
-  if (typeof window === 'undefined') {
-    return 'price'
-  }
-
-  const storedOddsFormat = window.localStorage.getItem(SPORTS_EVENT_ODDS_FORMAT_STORAGE_KEY)
-  const matchedOption = ODDS_FORMAT_OPTIONS.find(option => option.value === storedOddsFormat)
-  return matchedOption?.value ?? 'price'
-}
-
-function subscribeToOddsFormatStorage(listener: () => void) {
-  if (typeof window === 'undefined') {
-    return function unsubscribeFromOddsFormatStorage() {}
-  }
-
-  function handleStorage(event: StorageEvent) {
-    if (event.key && event.key !== SPORTS_EVENT_ODDS_FORMAT_STORAGE_KEY) {
-      return
-    }
-
-    listener()
-  }
-
-  window.addEventListener('storage', handleStorage)
-  return function unsubscribeFromOddsFormatStorage() {
-    window.removeEventListener('storage', handleStorage)
-  }
-}
-
-function getStoredOddsFormatClientSnapshot(): OddsFormat {
-  return resolveInitialOddsFormat()
-}
-
-function getStoredOddsFormatServerSnapshot(): OddsFormat {
-  return 'price'
-}
-
-function useOddsFormat() {
-  return useSyncExternalStore(
-    subscribeToOddsFormatStorage,
-    getStoredOddsFormatClientSnapshot,
-    getStoredOddsFormatServerSnapshot,
-  )
-}
-
-function areRecordValuesEqual<T extends string | null | undefined>(
-  left: Record<string, T>,
-  right: Record<string, T>,
-) {
-  const leftKeys = Object.keys(left)
-  const rightKeys = Object.keys(right)
-
-  if (leftKeys.length !== rightKeys.length) {
-    return false
-  }
-
-  return leftKeys.every(key => left[key] === right[key])
-}
-
-function parseRequestedOutcomeIndex(value: string | null | undefined) {
-  const rawValue = value?.trim() ?? ''
-  const parsed = Number.parseInt(rawValue, 10)
-  return Number.isFinite(parsed) ? parsed : null
-}
-
-function normalizeSportsMarketType(value: string | null | undefined) {
-  return value?.trim().toLowerCase() ?? ''
-}
-
-function resolveMoneylineButtonGridClass(buttonCount: number) {
-  if (buttonCount <= 1) {
-    return 'grid-cols-1'
-  }
-
-  if (buttonCount === 2) {
-    return 'grid-cols-2'
-  }
-
-  return 'grid-cols-3'
-}
-
-function parseEsportsSegmentDescriptor(market: SportsGamesCard['detailMarkets'][number] | null | undefined) {
-  if (!market) {
-    return null
-  }
-
-  const segmentMatch = [
-    market.sports_group_item_title,
-    market.short_title,
-    market.title,
-    market.slug,
-  ]
-    .filter((value): value is string => Boolean(value?.trim()))
-    .join(' ')
-    .match(/\b(map|game)\s*(\d+)\b/i)
-
-  if (!segmentMatch?.[1] || !segmentMatch[2]) {
-    return null
-  }
-
-  const parsed = Number.parseInt(segmentMatch[2], 10)
-  if (!Number.isFinite(parsed)) {
-    return null
-  }
-
-  return {
-    kind: segmentMatch[1].toLowerCase() === 'map' ? 'map' : 'game',
-    number: parsed,
-  } as const
-}
-
-function parseEsportsSegmentNumber(market: SportsGamesCard['detailMarkets'][number] | null | undefined) {
-  return parseEsportsSegmentDescriptor(market)?.number ?? null
-}
-
-function isSegmentedEsportsEventCard(card: SportsGamesCard, vertical: SportsVertical) {
-  return vertical === 'esports'
-    && card.detailMarkets.some(market => parseEsportsSegmentDescriptor(market) != null)
-}
-
-function resolveEsportsSegmentLabels(card: SportsGamesCard) {
-  const segmentKinds = new Set(
-    card.detailMarkets
-      .map(market => parseEsportsSegmentDescriptor(market)?.kind ?? null)
-      .filter((kind): kind is 'map' | 'game' => kind === 'map' || kind === 'game'),
-  )
-
-  if (segmentKinds.size === 1) {
-    const [kind] = Array.from(segmentKinds)
-    return kind === 'map'
-      ? { singular: 'Map', plural: 'Maps' }
-      : { singular: 'Game', plural: 'Games' }
-  }
-
-  return { singular: 'Game', plural: 'Games' }
-}
-
-function isSegmentedEsportsChildMoneylineMarket(market: SportsGamesCard['detailMarkets'][number] | null | undefined) {
-  return normalizeSportsMarketType(market?.sports_market_type) === 'child_moneyline'
-}
-
-function isSegmentedEsportsBinaryMarket(market: SportsGamesCard['detailMarkets'][number] | null | undefined) {
-  return parseEsportsSegmentNumber(market) != null && !isSegmentedEsportsChildMoneylineMarket(market)
-}
-
-function isSegmentedEsportsPrimaryMoneylineMarket(market: SportsGamesCard['detailMarkets'][number] | null | undefined) {
-  if (!market || isSegmentedEsportsChildMoneylineMarket(market)) {
-    return false
-  }
-
-  const normalizedType = normalizeSportsMarketType(market.sports_market_type)
-  if (normalizedType === 'moneyline') {
-    return true
-  }
-
-  const marketText = [
-    market.sports_group_item_title,
-    market.short_title,
-    market.title,
-  ]
-    .filter((value): value is string => Boolean(value?.trim()))
-    .join(' ')
-    .toLowerCase()
-
-  return marketText.includes('match winner') || marketText.includes('moneyline')
-}
-
-function resolveEsportsSegmentTabKey(mapNumber: number): EsportsLayoutTabKey {
-  return `segment-${mapNumber}`
-}
-
-function parseEsportsSegmentTabNumber(tabKey: EsportsLayoutTabKey) {
-  if (tabKey === 'series') {
-    return null
-  }
-
-  const parsed = Number.parseInt(tabKey.slice(8), 10)
-  return Number.isFinite(parsed) ? parsed : null
-}
-
-function resolveEsportsSegmentPanelTitle(markets: SportsGamesCard['detailMarkets']) {
-  const primaryMarket = markets[0]
-  if (!primaryMarket) {
-    return 'Market'
-  }
-
-  return primaryMarket.short_title?.trim()
-    || primaryMarket.sports_group_item_title?.trim()
-    || primaryMarket.title
-}
-
-function resolveEsportsSegmentPanelSortOrder(markets: SportsGamesCard['detailMarkets']) {
-  const normalizedType = normalizeSportsMarketType(markets[0]?.sports_market_type)
-  if (normalizedType === 'child_moneyline') {
-    return 0
-  }
-  if (normalizedType === 'spread' || normalizedType === 'map_handicap' || normalizedType.includes('handicap')) {
-    return 1
-  }
-  if (normalizedType === 'total' || normalizedType === 'totals' || normalizedType.includes('total')) {
-    return 2
-  }
-  if (normalizedType.endsWith('odd_even_total_kills')) {
-    return 3
-  }
-  if (normalizedType.endsWith('odd_even_total_rounds')) {
-    return 4
-  }
-  return 99
-}
-
-function useSportsSegmentNumberPicker({
-  options,
-  activeNumber,
-  onPick,
-}: {
-  options: SportsSegmentNumberPickerOption[]
-  activeNumber: number | null
-  onPick: (number: number) => void
-}) {
-  const scrollerRef = useRef<HTMLDivElement | null>(null)
-  const buttonRefsRef = useRef<Record<string, HTMLButtonElement | null>>({})
-  const [startSpacer, setStartSpacer] = useState(0)
-  const [endSpacer, setEndSpacer] = useState(0)
-
-  const activeOptionIndex = useMemo(
-    () => options.findIndex(option => option.number === activeNumber),
-    [activeNumber, options],
-  )
-
-  const pickOption = useCallback((optionIndex: number) => {
-    const option = options[optionIndex]
-    if (!option) {
-      return
-    }
-
-    onPick(option.number)
-  }, [onPick, options])
-
-  const handlePickPrevious = useCallback(() => {
-    if (activeOptionIndex <= 0) {
-      return
-    }
-
-    pickOption(activeOptionIndex - 1)
-  }, [activeOptionIndex, pickOption])
-
-  const handlePickNext = useCallback(() => {
-    if (activeOptionIndex < 0 || activeOptionIndex >= options.length - 1) {
-      return
-    }
-
-    pickOption(activeOptionIndex + 1)
-  }, [activeOptionIndex, options.length, pickOption])
-
-  const alignActiveOption = useCallback((behavior: ScrollBehavior = 'smooth') => {
-    if (activeOptionIndex < 0) {
-      return
-    }
-
-    const scroller = scrollerRef.current
-    if (!scroller) {
-      return
-    }
-
-    const activeOption = options[activeOptionIndex]
-    if (!activeOption) {
-      return
-    }
-
-    const activeButton = buttonRefsRef.current[activeOption.key]
-    if (!activeButton) {
-      return
-    }
-
-    const targetLeft = activeButton.offsetLeft - ((scroller.clientWidth - activeButton.offsetWidth) / 2)
-    scroller.scrollTo({
-      left: Math.max(0, targetLeft),
-      behavior,
-    })
-  }, [activeOptionIndex, options])
-
-  const updateSpacers = useCallback(() => {
-    const scroller = scrollerRef.current
-    if (!scroller || options.length === 0) {
-      setStartSpacer(0)
-      setEndSpacer(0)
-      return
-    }
-
-    const firstOptionKey = options[0]?.key
-    const lastOptionKey = options.at(-1)?.key
-    const firstButton = firstOptionKey ? buttonRefsRef.current[firstOptionKey] : null
-    const lastButton = lastOptionKey ? buttonRefsRef.current[lastOptionKey] : null
-    const fallbackButtonWidth = 40
-    const inferredButtonWidth = firstButton?.offsetWidth
-      ?? lastButton?.offsetWidth
-      ?? fallbackButtonWidth
-    const firstButtonWidth = firstButton?.offsetWidth ?? inferredButtonWidth
-    const lastButtonWidth = lastButton?.offsetWidth ?? inferredButtonWidth
-    const viewportWidth = scroller.clientWidth
-    const scrollerStyles = window.getComputedStyle(scroller)
-    const gapWidth = Number.parseFloat(scrollerStyles.columnGap || scrollerStyles.gap || '0') || 0
-    const startSpacerWidth = Math.max(0, viewportWidth / 2 - firstButtonWidth / 2 - gapWidth)
-    const endSpacerWidth = Math.max(0, viewportWidth / 2 - lastButtonWidth / 2 - gapWidth)
-
-    setStartSpacer(startSpacerWidth)
-    setEndSpacer(endSpacerWidth)
-  }, [options])
-
-  useEffect(function alignOnActiveOptionChange() {
-    if (activeOptionIndex < 0) {
-      return
-    }
-
-    alignActiveOption('auto')
-  }, [activeOptionIndex, alignActiveOption, endSpacer, startSpacer])
-
-  useEffect(function scheduleSpacerAndAlignmentUpdate() {
-    if (options.length <= 1) {
-      return
-    }
-
-    const frame = window.requestAnimationFrame(() => {
-      updateSpacers()
-      alignActiveOption('auto')
-    })
-
-    return function cancelScheduledSpacerAndAlignmentUpdate() {
-      window.cancelAnimationFrame(frame)
-    }
-  }, [alignActiveOption, options.length, updateSpacers])
-
-  useEffect(function observeScrollerResizeForSpacerUpdate() {
-    const scrollerElement = scrollerRef.current
-    if (options.length <= 1 || !scrollerElement) {
-      return
-    }
-
-    updateSpacers()
-
-    if (typeof ResizeObserver === 'undefined') {
-      window.addEventListener('resize', updateSpacers)
-      return function removeResizeListener() {
-        window.removeEventListener('resize', updateSpacers)
-      }
-    }
-
-    const observer = new ResizeObserver(updateSpacers)
-    observer.observe(scrollerElement)
-    return function disconnectResizeObserver() {
-      observer.disconnect()
-    }
-  }, [options.length, updateSpacers])
-
-  return {
-    scrollerRef,
-    buttonRefsRef,
-    startSpacer,
-    endSpacer,
-    activeOptionIndex,
-    pickOption,
-    handlePickPrevious,
-    handlePickNext,
-  }
-}
-
-function SportsSegmentNumberPicker({
-  options,
-  activeNumber,
-  segmentLabel,
-  onPick,
-}: {
-  options: SportsSegmentNumberPickerOption[]
-  activeNumber: number | null
-  segmentLabel: string
-  onPick: (number: number) => void
-}) {
-  const {
-    scrollerRef,
-    buttonRefsRef,
-    startSpacer,
-    endSpacer,
-    activeOptionIndex,
-    pickOption,
-    handlePickPrevious,
-    handlePickNext,
-  } = useSportsSegmentNumberPicker({ options, activeNumber, onPick })
-
-  if (options.length <= 1) {
-    return null
-  }
-
-  return (
-    <div className="bg-card px-2.5 pb-2">
-      <div className="-mx-2.5 border-t border-border/70" />
-      <div className="pt-2">
-        <div className="mt-0.5 flex items-center gap-2">
-          <button
-            type="button"
-            onClick={handlePickPrevious}
-            disabled={activeOptionIndex <= 0}
-            className={cn(
-              `
-                inline-flex size-7 items-center justify-center rounded-sm text-muted-foreground transition-colors
-                focus-visible:ring-1 focus-visible:ring-ring focus-visible:outline-none
-              `,
-              activeOptionIndex > 0
-                ? 'cursor-pointer hover:bg-muted/70 hover:text-foreground'
-                : 'cursor-not-allowed opacity-40',
-            )}
-            aria-label={`Previous ${segmentLabel.toLowerCase()}`}
-          >
-            <ChevronLeftIcon className="size-4.5" />
-          </button>
-
-          <div className="relative min-w-0 flex-1">
-            <span
-              aria-hidden
-              className="
-                pointer-events-none absolute -top-2 left-1/2 h-2 w-3 -translate-x-1/2 bg-primary
-                [clip-path:polygon(50%_100%,0_0,100%_0)]
-              "
-            />
-
-            <div
-              ref={scrollerRef}
-              className={`
-                flex min-w-0 snap-x snap-mandatory items-center gap-2 overflow-x-auto scroll-smooth
-                [scrollbar-width:none]
-                [&::-webkit-scrollbar]:hidden
-              `}
-            >
-              <span aria-hidden className="shrink-0" style={{ width: startSpacer }} />
-              {options.map((option, index) => (
-                <button
-                  key={option.key}
-                  type="button"
-                  onClick={() => pickOption(index)}
-                  ref={(node) => {
-                    buttonRefsRef.current[option.key] = node
-                  }}
-                  className={cn(
-                    `w-10 shrink-0 snap-center text-center text-sm font-medium text-muted-foreground transition-colors`,
-                    index === activeOptionIndex
-                      ? 'text-base font-semibold text-foreground'
-                      : 'hover:text-foreground/80',
-                  )}
-                  aria-label={`${segmentLabel} ${option.number}`}
-                >
-                  {option.label}
-                </button>
-              ))}
-              <span aria-hidden className="shrink-0" style={{ width: endSpacer }} />
-            </div>
-          </div>
-
-          <button
-            type="button"
-            onClick={handlePickNext}
-            disabled={activeOptionIndex < 0 || activeOptionIndex >= options.length - 1}
-            className={cn(
-              `
-                inline-flex size-7 items-center justify-center rounded-sm text-muted-foreground transition-colors
-                focus-visible:ring-1 focus-visible:ring-ring focus-visible:outline-none
-              `,
-              activeOptionIndex >= 0 && activeOptionIndex < options.length - 1
-                ? 'cursor-pointer hover:bg-muted/70 hover:text-foreground'
-                : 'cursor-not-allowed opacity-40',
-            )}
-            aria-label={`Next ${segmentLabel.toLowerCase()}`}
-          >
-            <ChevronRightIcon className="size-4.5" />
-          </button>
-        </div>
-      </div>
-    </div>
-  )
-}
-
-function useSportsEventQuerySync(onSelectionChange: (selection: SportsEventQuerySelection) => void) {
-  const searchParams = useSearchParams()
-
-  useEffect(function syncQuerySelectionFromSearchParams() {
-    onSelectionChange({
-      conditionId: searchParams.get('conditionId')?.trim() ?? null,
-      outcomeIndex: parseRequestedOutcomeIndex(searchParams.get('outcomeIndex')),
-    })
-
-    return function noopQuerySelectionSyncCleanup() {}
-  }, [onSelectionChange, searchParams])
-}
-
-function SportsEventQuerySync({
-  onSelectionChange,
-}: {
-  onSelectionChange: (selection: SportsEventQuerySelection) => void
-}) {
-  useSportsEventQuerySync(onSelectionChange)
-
-  return null
-}
-
-function resolvePositionShares(position: UserPosition) {
-  const totalShares = typeof position.total_shares === 'number' ? position.total_shares : Number(position.size ?? 0)
-  return Number.isFinite(totalShares) ? totalShares : 0
-}
-
-function resolveOutcomeIndexFromPosition(position: UserPosition) {
-  if (position.outcome_index === OUTCOME_INDEX.YES || position.outcome_index === OUTCOME_INDEX.NO) {
-    return position.outcome_index
-  }
-
-  const normalizedOutcome = position.outcome_text?.trim().toLowerCase()
-  if (normalizedOutcome === 'no') {
-    return OUTCOME_INDEX.NO
-  }
-  if (normalizedOutcome === 'yes') {
-    return OUTCOME_INDEX.YES
-  }
-  return null
-}
-
-function resolveIndexSetFromOutcomeIndex(outcomeIndex: number | null | undefined) {
-  if (outcomeIndex === OUTCOME_INDEX.YES) {
-    return 1
-  }
-  if (outcomeIndex === OUTCOME_INDEX.NO) {
-    return 2
-  }
-  return null
-}
-
-function resolveTeamShortLabel(team: SportsGamesCard['teams'][number] | null | undefined) {
-  const abbreviation = team?.abbreviation?.trim()
-  if (abbreviation) {
-    return abbreviation.toUpperCase()
-  }
-
-  const name = team?.name?.trim()
-  if (!name) {
-    return '—'
-  }
-
-  const initials = name
-    .split(/\s+/)
-    .filter(Boolean)
-    .map(part => part[0]?.toUpperCase() ?? '')
-    .join('')
-    .slice(0, 3)
-
-  return initials || name.slice(0, 3).toUpperCase()
-}
-
-const FULL_COMPETITOR_NAME_HERO_LABEL_SPORT_SLUGS = new Set([
-  'boxing',
-  'chess',
-  'f1',
-  'golf',
-  'mma',
-  'tennis',
-  'ufc',
-  'wtt-mens-singles',
-  'zuffa',
-])
-
-function shouldUseFullCompetitorHeroLabels(sportSlug: string | null | undefined) {
-  return FULL_COMPETITOR_NAME_HERO_LABEL_SPORT_SLUGS.has(
-    normalizeComparableToken(sportSlug),
-  )
-}
-
-function shouldUseFullScoreboardHeroLabels({
-  sportSlug,
-  vertical,
-}: {
-  sportSlug: string | null | undefined
-  vertical: SportsVertical
-}) {
-  return vertical === 'esports' || shouldUseFullCompetitorHeroLabels(sportSlug)
-}
-
-function parseSportsScore(value: string | null | undefined) {
-  const trimmed = value?.trim()
-  if (!trimmed) {
-    return null
-  }
-
-  const match = trimmed.match(/(\d+)\D+(\d+)/)
-  if (!match) {
-    return null
-  }
-
-  const team1 = Number.parseInt(match[1] ?? '', 10)
-  const team2 = Number.parseInt(match[2] ?? '', 10)
-  if (!Number.isFinite(team1) || !Number.isFinite(team2)) {
-    return null
-  }
-
-  return { team1, team2 }
-}
-
-function resolveRelatedTeamOdds(card: SportsGamesCard) {
-  const moneylineButtons = card.buttons.filter(button => button.marketType === 'moneyline')
-  const team1Button = moneylineButtons.find(button => button.tone === 'team1') ?? moneylineButtons[0] ?? null
-  const team2Button = moneylineButtons.find(button => button.tone === 'team2')
-    ?? moneylineButtons.find(button => button.key !== team1Button?.key)
-    ?? null
-
-  return {
-    team1Cents: team1Button?.cents ?? null,
-    team2Cents: team2Button?.cents ?? null,
-  }
-}
-
-function formatRelatedOddsLabel(cents: number | null) {
-  if (typeof cents !== 'number' || !Number.isFinite(cents)) {
-    return '—'
-  }
-  return `${cents}¢`
-}
-
-function normalizeComparableToken(value: string | null | undefined) {
-  return value
-    ?.normalize('NFKD')
-    .replace(/[\u0300-\u036F]/g, '')
-    .toLowerCase()
-    .replace(/[^a-z0-9]+/g, '')
-    .trim()
-    ?? ''
-}
-
-function tokenizeComparableText(value: string | null | undefined) {
-  return value
-    ?.normalize('NFKD')
-    .replace(/[\u0300-\u036F]/g, '')
-    .toLowerCase()
-    .split(/[^a-z0-9]+/g)
-    .filter(Boolean)
-    ?? []
-}
-
-function doesTextMatchSportsTeam(
-  value: string | null | undefined,
-  team: SportsGamesCard['teams'][number] | null | undefined,
-) {
-  if (!value || !team) {
-    return false
-  }
-
-  const normalizedValue = normalizeComparableToken(value)
-  if (!normalizedValue) {
-    return false
-  }
-
-  const normalizedTeamName = normalizeComparableToken(team.name)
-  if (normalizedTeamName && normalizedValue.includes(normalizedTeamName)) {
-    return true
-  }
-
-  const normalizedTeamAbbreviation = normalizeComparableToken(team.abbreviation)
-  if (!normalizedTeamAbbreviation) {
-    return false
-  }
-
-  return new Set(tokenizeComparableText(value)).has(normalizedTeamAbbreviation)
-}
-
-function resolveSegmentedEsportsButtonMarketType(
-  market: SportsGamesCard['detailMarkets'][number],
-): SportsGamesButton['marketType'] {
-  const normalizedType = normalizeSportsMarketType(market.sports_market_type)
-
-  if (isSegmentedEsportsChildMoneylineMarket(market) || isSegmentedEsportsPrimaryMoneylineMarket(market)) {
-    return 'moneyline'
-  }
-
-  if (
-    normalizedType === 'spread'
-    || normalizedType === 'map_handicap'
-    || normalizedType.includes('handicap')
-  ) {
-    return 'spread'
-  }
-
-  if (
-    normalizedType === 'total'
-    || normalizedType === 'totals'
-    || normalizedType.includes('total')
-  ) {
-    return 'total'
-  }
-
-  if (
-    normalizedType === 'btts'
-    || normalizedType.includes('both_teams_to_score')
-    || normalizedType.includes('both teams to score')
-  ) {
-    return 'btts'
-  }
-
-  return 'binary'
-}
-
-function buildSegmentedEsportsButtonsFromOutcomes(
-  card: SportsGamesCard,
-  market: SportsGamesCard['detailMarkets'][number],
-) {
-  const team1 = card.teams[0] ?? null
-  const team2 = card.teams[1] ?? null
-  const marketType = resolveSegmentedEsportsButtonMarketType(market)
-
-  return [...market.outcomes]
-    .sort((left, right) => left.outcome_index - right.outcome_index)
-    .map((outcome) => {
-      const outcomeText = outcome.outcome_text?.trim() ?? ''
-      const normalizedOutcomeText = normalizeComparableToken(outcomeText)
-      const matchedTeam = [team1, team2].find(team => doesTextMatchSportsTeam(outcomeText, team)) ?? null
-
-      let label = outcomeText.toUpperCase() || 'MARKET'
-      let color: string | null = null
-      let tone: SportsGamesButton['tone'] = 'neutral'
-
-      if (matchedTeam === team1) {
-        label = resolveTeamShortLabel(team1)
-        color = team1?.color ?? null
-        tone = 'team1'
-      }
-      else if (matchedTeam === team2) {
-        label = resolveTeamShortLabel(team2)
-        color = team2?.color ?? null
-        tone = 'team2'
-      }
-      else if (normalizedOutcomeText.includes('draw')) {
-        label = 'DRAW'
-        tone = 'draw'
-      }
-      else if (normalizedOutcomeText === 'yes') {
-        label = 'YES'
-        tone = 'over'
-      }
-      else if (normalizedOutcomeText === 'no') {
-        label = 'NO'
-        tone = 'under'
-      }
-      else if (normalizedOutcomeText === 'over') {
-        label = 'OVER'
-        tone = 'over'
-      }
-      else if (normalizedOutcomeText === 'under') {
-        label = 'UNDER'
-        tone = 'under'
-      }
-
-      const fallbackIsNoOutcome = normalizedOutcomeText === 'no'
-
-      return {
-        key: `${market.condition_id}:${outcome.outcome_index}`,
-        conditionId: market.condition_id,
-        outcomeIndex: outcome.outcome_index,
-        fallbackIsNoOutcome,
-        label,
-        cents: resolveOutcomeSelectionPriceCents(market, outcome, {
-          side: ORDER_SIDE.BUY,
-          fallbackIsNoOutcome,
-        }) ?? 50,
-        color,
-        marketType,
-        tone,
-      } satisfies SportsGamesButton
-    })
-}
-
-function shouldNormalizeSegmentedEsportsMarketButtons(
-  market: SportsGamesCard['detailMarkets'][number],
-  currentButtons: SportsGamesButton[],
-) {
-  if (parseEsportsSegmentNumber(market) == null || market.outcomes.length <= 1) {
-    return false
-  }
-
-  if (currentButtons.length !== market.outcomes.length) {
-    return true
-  }
-
-  const expectedMarketType = resolveSegmentedEsportsButtonMarketType(market)
-  if (currentButtons.some(button => button.marketType !== expectedMarketType)) {
-    return true
-  }
-
-  const currentOutcomeIndexes = new Set(currentButtons.map(button => button.outcomeIndex))
-  return market.outcomes.some(outcome => !currentOutcomeIndexes.has(outcome.outcome_index))
-}
-
-function resolveNormalizedSegmentedEsportsCard(
-  card: SportsGamesCard,
-  vertical: SportsVertical,
-) {
-  if (!isSegmentedEsportsEventCard(card, vertical)) {
-    return card
-  }
-
-  const buttonsByConditionId = new Map<string, SportsGamesButton[]>()
-  card.buttons.forEach((button) => {
-    const currentButtons = buttonsByConditionId.get(button.conditionId)
-    if (currentButtons) {
-      currentButtons.push(button)
-      return
-    }
-
-    buttonsByConditionId.set(button.conditionId, [button])
-  })
-
-  const rebuiltButtonsByConditionId = new Map<string, SportsGamesButton[]>()
-  card.detailMarkets.forEach((market) => {
-    const currentButtons = buttonsByConditionId.get(market.condition_id) ?? []
-    if (!shouldNormalizeSegmentedEsportsMarketButtons(market, currentButtons)) {
-      return
-    }
-
-    const rebuiltButtons = buildSegmentedEsportsButtonsFromOutcomes(card, market)
-    if (rebuiltButtons.length > 0) {
-      rebuiltButtonsByConditionId.set(market.condition_id, rebuiltButtons)
-    }
-  })
-
-  if (rebuiltButtonsByConditionId.size === 0) {
-    return card
-  }
-
-  const nextButtons: SportsGamesButton[] = []
-  const insertedConditionIds = new Set<string>()
-
-  card.buttons.forEach((button) => {
-    const rebuiltButtons = rebuiltButtonsByConditionId.get(button.conditionId)
-    if (rebuiltButtons) {
-      if (!insertedConditionIds.has(button.conditionId)) {
-        nextButtons.push(...rebuiltButtons)
-        insertedConditionIds.add(button.conditionId)
-      }
-      return
-    }
-
-    nextButtons.push(button)
-  })
-
-  card.detailMarkets.forEach((market) => {
-    if (insertedConditionIds.has(market.condition_id)) {
-      return
-    }
-
-    const rebuiltButtons = rebuiltButtonsByConditionId.get(market.condition_id)
-    if (!rebuiltButtons) {
-      return
-    }
-
-    nextButtons.push(...rebuiltButtons)
-    insertedConditionIds.add(market.condition_id)
-  })
-
-  return {
-    ...card,
-    buttons: nextButtons,
-  }
-}
-
-function resolveTeamByTone(card: SportsGamesCard, tone: SportsGamesButton['tone']) {
-  if (tone === 'team1') {
-    return card.teams[0] ?? null
-  }
-  if (tone === 'team2') {
-    return card.teams[1] ?? null
-  }
-  return null
-}
-
-function resolveRedeemOptionLabel(
-  card: SportsGamesCard,
-  market: SportsGamesCard['detailMarkets'][number],
-  button: SportsGamesButton,
-) {
-  const rawLabel = button.label?.trim() ?? ''
-  const team = resolveTeamByTone(card, button.tone)
-
-  if (team?.name) {
-    const firstToken = rawLabel.split(/\s+/)[0] ?? ''
-    const normalizedFirstToken = normalizeComparableToken(firstToken)
-    const normalizedTeamAbbreviation = normalizeComparableToken(team.abbreviation)
-    const normalizedTeamName = normalizeComparableToken(team.name)
-
-    if (normalizedFirstToken && (
-      (normalizedTeamAbbreviation && normalizedFirstToken === normalizedTeamAbbreviation)
-      || (normalizedTeamName && normalizedTeamName.startsWith(normalizedFirstToken))
-    )) {
-      return `${team.name}${rawLabel.slice(firstToken.length)}`
-    }
-
-    return team.name
-  }
-
-  if (button.tone === 'draw') {
-    return 'Draw'
-  }
-
-  if (button.tone === 'over') {
-    return rawLabel || 'Over'
-  }
-
-  if (button.tone === 'under') {
-    return rawLabel || 'Under'
-  }
-
-  return market.sports_group_item_title?.trim()
-    || market.short_title?.trim()
-    || market.title
-    || rawLabel
-}
-
-function normalizeHexColor(value: string | null | undefined) {
-  if (!value) {
-    return null
-  }
-
-  let normalized = value.trim()
-  if (!normalized) {
-    return null
-  }
-
-  if (!normalized.startsWith('#')) {
-    normalized = `#${normalized}`
-  }
-
-  if (/^#[0-9A-F]{3}$/i.test(normalized) || /^#[0-9A-F]{6}$/i.test(normalized)) {
-    return normalized
-  }
-
-  return null
-}
-
-function resolveHexToRgbComponents(value: string) {
-  const hex = value.replace('#', '')
-  const expandedHex = hex.length === 3
-    ? hex.split('').map(char => `${char}${char}`).join('')
-    : hex
-
-  const red = Number.parseInt(expandedHex.slice(0, 2), 16)
-  const green = Number.parseInt(expandedHex.slice(2, 4), 16)
-  const blue = Number.parseInt(expandedHex.slice(4, 6), 16)
-  if ([red, green, blue].some(component => Number.isNaN(component))) {
-    return null
-  }
-
-  return `${red} ${green} ${blue}`
-}
-
-function resolveRedeemTagAccent(
-  button: SportsGamesButton | null,
-  outcomeIndex: number | null,
-) {
-  const normalizedTeamColor = normalizeHexColor(button?.color)
-  if (
-    button
-    && (button.tone === 'team1' || button.tone === 'team2')
-    && normalizedTeamColor
-  ) {
-    const rgbComponents = resolveHexToRgbComponents(normalizedTeamColor)
-    const readableTeamColor = ensureReadableTextColorOnDark(normalizedTeamColor)
-    return {
-      badgeClassName: '',
-      badgeStyle: {
-        color: readableTeamColor ?? normalizedTeamColor,
-        backgroundColor: rgbComponents ? `rgb(${rgbComponents} / 0.10)` : undefined,
-      } as const,
-    }
-  }
-
-  if ((button && button.tone === 'over') || outcomeIndex === OUTCOME_INDEX.YES) {
-    return {
-      badgeClassName: 'bg-yes/10 text-yes',
-      badgeStyle: undefined,
-    }
-  }
-
-  if ((button && button.tone === 'under') || outcomeIndex === OUTCOME_INDEX.NO) {
-    return {
-      badgeClassName: 'bg-no/10 text-no',
-      badgeStyle: undefined,
-    }
-  }
-
-  return {
-    badgeClassName: 'bg-muted/60 text-muted-foreground',
-    badgeStyle: undefined,
-  }
-}
-
-function useSportsEventShareButton(event: SportsGamesCard['event']) {
-  const [shareSuccess, setShareSuccess] = useState(false)
-  const debugPayload = useMemo(() => {
-    return {
-      event: {
-        id: event.id,
-        slug: event.slug,
-        title: event.title,
-      },
-      markets: (event.markets ?? []).map(market => ({
-        slug: market.slug,
-        condition_id: market.condition_id,
-        question_id: market.question_id,
-        metadata_hash: market.condition?.metadata_hash ?? null,
-        short_title: market.short_title ?? null,
-        title: market.title,
-        outcomes: market.outcomes.map(outcome => ({
-          outcome_index: outcome.outcome_index,
-          outcome_text: outcome.outcome_text,
-          token_id: outcome.token_id,
-        })),
-      })),
-    }
-  }, [event.id, event.markets, event.slug, event.title])
-
-  const handleDebugCopy = useCallback(async () => {
-    try {
-      await navigator.clipboard.writeText(JSON.stringify(debugPayload, null, 2))
-    }
-    catch {
-      // noop
-    }
-  }, [debugPayload])
-
-  const maybeHandleDebugCopy = useCallback((event: React.MouseEvent) => {
-    if (!event.altKey) {
-      return false
-    }
-
-    event.preventDefault()
-    event.stopPropagation()
-    void handleDebugCopy()
-    return true
-  }, [handleDebugCopy])
-
-  return { shareSuccess, setShareSuccess, maybeHandleDebugCopy }
-}
-
-function SportsEventShareButton({ event }: { event: SportsGamesCard['event'] }) {
-  const user = useUser()
-  const affiliateCode = user?.affiliate_code?.trim() ?? ''
-  const { shareSuccess, setShareSuccess, maybeHandleDebugCopy } = useSportsEventShareButton(event)
-
-  async function handleShare() {
-    try {
-      const url = new URL(window.location.href)
-      if (affiliateCode) {
-        url.searchParams.set('r', affiliateCode)
-      }
-      await navigator.clipboard.writeText(url.toString())
-      setShareSuccess(true)
-      window.setTimeout(setShareSuccess, 2000, false)
-    }
-    catch {
-      // noop
-    }
-  }
-
-  return (
-    <Button
-      type="button"
-      variant="ghost"
-      size="icon"
-      className={cn(headerIconButtonClass, 'size-auto p-0')}
-      aria-label="Copy event link"
-      onClick={(event) => {
-        if (maybeHandleDebugCopy(event)) {
-          return
-        }
-        void handleShare()
-      }}
-    >
-      {shareSuccess
-        ? <CheckIcon className="size-4 text-primary" />
-        : <ShareIcon className="size-4" />}
-    </Button>
-  )
-}
-
-function normalizeLivestreamUrl(value: string | null | undefined) {
-  const trimmed = value?.trim()
-  if (!trimmed) {
-    return null
-  }
-
-  try {
-    const parsed = new URL(trimmed)
-    if (parsed.protocol !== 'http:' && parsed.protocol !== 'https:') {
-      return null
-    }
-    return parsed.toString()
-  }
-  catch {
-    return null
-  }
-}
-
-function SportsEventLiveStatusIcon({ className, muted = false }: { className?: string, muted?: boolean }) {
-  if (muted) {
-    return (
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        width="18"
-        height="18"
-        viewBox="0 0 18 18"
-        className={cn(className, 'text-muted-foreground')}
-        fill="none"
-      >
-        <g stroke="currentColor" strokeLinecap="round" strokeLinejoin="round" strokeWidth="1.5">
-          <path d="M5.641,12.359c-1.855-1.855-1.855-4.863,0-6.718" />
-          <path d="M3.52,14.48C.493,11.454,.493,6.546,3.52,3.52" />
-          <circle cx="9" cy="9" r="1.75" fill="none" stroke="currentColor" />
-          <path d="M12.359,12.359c1.855-1.855,1.855-4.863,0-6.718" />
-          <path d="M14.48,14.48c3.027-3.027,3.027-7.934,0-10.96" />
-        </g>
-      </svg>
-    )
-  }
-
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="18"
-      height="18"
-      viewBox="0 0 18 18"
-      className={cn(className, 'text-red-500')}
-      fill="none"
-    >
-      <g stroke="currentColor" strokeLinecap="round" strokeLinejoin="round" strokeWidth="1.5">
-        <path d="M5.641,12.359c-1.855-1.855-1.855-4.863,0-6.718" opacity="0.24">
-          <animate
-            attributeName="opacity"
-            values="0.24;1;1;0.24;0.24"
-            keyTimes="0;0.28;0.56;0.84;1"
-            dur="1.8s"
-            repeatCount="indefinite"
-          />
-        </path>
-        <path d="M3.52,14.48C.493,11.454,.493,6.546,3.52,3.52" opacity="0.14">
-          <animate
-            attributeName="opacity"
-            values="0.14;0.14;0.92;0.92;0.14;0.14"
-            keyTimes="0;0.4;0.58;0.78;0.92;1"
-            dur="1.8s"
-            repeatCount="indefinite"
-          />
-        </path>
-        <circle cx="9" cy="9" r="1.75" fill="none" stroke="currentColor" />
-        <path d="M12.359,12.359c1.855-1.855,1.855-4.863,0-6.718" opacity="0.24">
-          <animate
-            attributeName="opacity"
-            values="0.24;1;1;0.24;0.24"
-            keyTimes="0;0.28;0.56;0.84;1"
-            dur="1.8s"
-            repeatCount="indefinite"
-          />
-        </path>
-        <path d="M14.48,14.48c3.027-3.027,3.027-7.934,0-10.96" opacity="0.14">
-          <animate
-            attributeName="opacity"
-            values="0.14;0.14;0.92;0.92;0.14;0.14"
-            keyTimes="0;0.4;0.58;0.78;0.92;1"
-            dur="1.8s"
-            repeatCount="indefinite"
-          />
-        </path>
-      </g>
-    </svg>
-  )
-}
-
-function sortSectionButtons(sectionKey: EventSectionKey, buttons: SportsGamesButton[]) {
-  if (sectionKey === 'spread') {
-    const order: Record<SportsGamesButton['tone'], number> = {
-      team1: 0,
-      team2: 1,
-      draw: 2,
-      over: 3,
-      under: 4,
-      neutral: 5,
-    }
-
-    return [...buttons].sort((a, b) => (order[a.tone] ?? 99) - (order[b.tone] ?? 99))
-  }
-
-  if (sectionKey === 'total' || sectionKey === 'btts') {
-    const order: Record<SportsGamesButton['tone'], number> = {
-      over: 0,
-      under: 1,
-      team1: 2,
-      team2: 3,
-      draw: 4,
-      neutral: 5,
-    }
-
-    return [...buttons].sort((a, b) => (order[a.tone] ?? 99) - (order[b.tone] ?? 99))
-  }
-
-  return buttons
-}
-
-function sortAuxiliaryButtons(buttons: SportsGamesButton[]) {
-  const order: Record<SportsGamesButton['tone'], number> = {
-    team1: 0,
-    draw: 1,
-    team2: 2,
-    over: 3,
-    under: 4,
-    neutral: 5,
-  }
-
-  return [...buttons].sort((a, b) => (order[a.tone] ?? 99) - (order[b.tone] ?? 99))
-}
-
-function isEventSectionKey(value: SportsGamesButton['marketType']): value is EventSectionKey {
-  return value === 'moneyline' || value === 'spread' || value === 'total' || value === 'btts'
-}
-
-function resolveEventSectionKeyForButton(
-  button: SportsGamesButton | null | undefined,
-  market: SportsGamesCard['detailMarkets'][number] | null | undefined,
-): EventSectionKey | null {
-  if (market) {
-    if (isSegmentedEsportsChildMoneylineMarket(market) || isSegmentedEsportsBinaryMarket(market)) {
-      return null
-    }
-
-    const normalizedType = normalizeSportsMarketType(market.sports_market_type)
-    if (isSegmentedEsportsPrimaryMoneylineMarket(market) || normalizedType === 'moneyline') {
-      return 'moneyline'
-    }
-
-    if (
-      normalizedType === 'spread'
-      || normalizedType === 'map_handicap'
-      || normalizedType.includes('handicap')
-    ) {
-      return 'spread'
-    }
-
-    if (
-      normalizedType === 'total'
-      || normalizedType === 'totals'
-      || normalizedType.includes('total')
-    ) {
-      return 'total'
-    }
-
-    if (
-      normalizedType === 'btts'
-      || normalizedType.includes('both_teams_to_score')
-      || normalizedType.includes('both teams to score')
-    ) {
-      return 'btts'
-    }
-  }
-
-  if (!button) {
-    return null
-  }
-
-  if (button.marketType === 'binary') {
-    return 'moneyline'
-  }
-
-  return isEventSectionKey(button.marketType) ? button.marketType : null
-}
-
-function resolveMarketViewCardBySlug(
-  marketViewCards: SportsGamesCardMarketView[],
-  marketSlug: string | null,
-) {
-  if (!marketSlug) {
-    return null
-  }
-
-  return marketViewCards.find(view =>
-    view.card.detailMarkets.some(market => market.slug === marketSlug),
-  ) ?? null
-}
-
-function dedupeAuxiliaryButtons(buttons: SportsGamesButton[]) {
-  const byKey = new Map<string, SportsGamesButton>()
-  buttons.forEach((button) => {
-    byKey.set(button.key, button)
-  })
-  return Array.from(byKey.values())
-}
-
-function resolveAuxiliaryPanelCreatedAt(markets: SportsGamesCard['detailMarkets']) {
-  return markets.reduce<number>((earliestTimestamp, market) => {
-    const timestamp = Date.parse(market.created_at)
-    if (!Number.isFinite(timestamp)) {
-      return earliestTimestamp
-    }
-
-    return Math.min(earliestTimestamp, timestamp)
-  }, Number.POSITIVE_INFINITY)
-}
-
-function resolveAuxiliaryPanelThreshold(markets: SportsGamesCard['detailMarkets']) {
-  return markets.reduce<number>((lowestThreshold, market) => {
-    const threshold = Number(market.sports_group_item_threshold)
-    if (!Number.isFinite(threshold)) {
-      return lowestThreshold
-    }
-
-    return Math.min(lowestThreshold, threshold)
-  }, Number.POSITIVE_INFINITY)
-}
-
-function SportsEventRelatedGames({
-  cards,
-  sportSlug,
-  sportLabel,
-  locale,
-  vertical,
-}: {
-  cards: SportsGamesCard[]
-  sportSlug: string
-  sportLabel: string
-  locale: string
-  vertical: SportsVertical
-}) {
-  const verticalConfig = getSportsVerticalConfig(vertical)
-  const dateTimeFormatter = useMemo(
-    () => new Intl.DateTimeFormat(locale, {
-      month: 'short',
-      day: 'numeric',
-      hour: 'numeric',
-      minute: '2-digit',
-    }),
-    [locale],
-  )
-
-  if (cards.length === 0) {
-    return null
-  }
-
-  return (
-    <div className="grid gap-2.5">
-      <p className="text-sm font-normal text-muted-foreground">
-        {'More '}
-        <AppLink href={`${verticalConfig.basePath}/${sportSlug}/games`} className="underline-offset-2 hover:underline">
-          {sportLabel}
-        </AppLink>
-        {' Games'}
-      </p>
-
-      <div className="grid gap-2">
-        {cards.map((relatedCard) => {
-          const startTime = relatedCard.startTime ? new Date(relatedCard.startTime) : null
-          const hasValidStartTime = Boolean(startTime && !Number.isNaN(startTime.getTime()))
-          const topLineDate = hasValidStartTime ? dateTimeFormatter.format(startTime as Date) : 'Date TBD'
-          const { team1Cents, team2Cents } = resolveRelatedTeamOdds(relatedCard)
-          const team1 = relatedCard.teams[0] ?? null
-          const team2 = relatedCard.teams[1] ?? null
-
-          return (
-            <AppLink
-              intentPrefetch
-              key={relatedCard.id}
-              href={relatedCard.eventHref}
-              className={cn(`block rounded-xl px-3 py-2.5 transition-colors hover:bg-card`)}
-            >
-              <p className="mb-2 text-xs font-normal text-muted-foreground">
-                {topLineDate}
-                <span className="mx-2 inline-block">·</span>
-                {formatVolume(relatedCard.volume)}
-                {' '}
-                Vol.
-              </p>
-
-              <div className="space-y-2.5">
-                <div className="flex items-center justify-between gap-2">
-                  <span className="flex min-w-0 items-center gap-2">
-                    <span className="flex size-6 shrink-0 items-center justify-center">
-                      {team1?.logoUrl
-                        ? (
-                            <Image
-                              src={team1.logoUrl}
-                              alt={`${team1.name} logo`}
-                              width={24}
-                              height={24}
-                              sizes="24px"
-                              className="size-full object-contain object-center"
-                            />
-                          )
-                        : (
-                            <span className="text-2xs font-semibold text-muted-foreground">
-                              {team1?.abbreviation?.slice(0, 1)?.toUpperCase() ?? '—'}
-                            </span>
-                          )}
-                    </span>
-                    <span className="truncate text-xs font-normal text-foreground">
-                      {team1?.name ?? '—'}
-                    </span>
-                  </span>
-                  <span className="shrink-0 text-sm font-normal text-muted-foreground">
-                    {formatRelatedOddsLabel(team1Cents)}
-                  </span>
-                </div>
-
-                <div className="flex items-center justify-between gap-2">
-                  <span className="flex min-w-0 items-center gap-2">
-                    <span className="flex size-6 shrink-0 items-center justify-center">
-                      {team2?.logoUrl
-                        ? (
-                            <Image
-                              src={team2.logoUrl}
-                              alt={`${team2.name} logo`}
-                              width={24}
-                              height={24}
-                              sizes="24px"
-                              className="size-full object-contain object-center"
-                            />
-                          )
-                        : (
-                            <span className="text-2xs font-semibold text-muted-foreground">
-                              {team2?.abbreviation?.slice(0, 1)?.toUpperCase() ?? '—'}
-                            </span>
-                          )}
-                    </span>
-                    <span className="truncate text-xs font-normal text-foreground">
-                      {team2?.name ?? '—'}
-                    </span>
-                  </span>
-                  <span className="shrink-0 text-sm font-normal text-muted-foreground">
-                    {formatRelatedOddsLabel(team2Cents)}
-                  </span>
-                </div>
-              </div>
-            </AppLink>
-          )
-        })}
-      </div>
-    </div>
-  )
-}
-
-function useQuerySelection() {
-  const [querySelection, setQuerySelection] = useState<SportsEventQuerySelection>(EMPTY_QUERY_SELECTION)
-
-  const handleQuerySelectionChange = useCallback((nextSelection: SportsEventQuerySelection) => {
-    setQuerySelection((current) => {
-      if (
-        current.conditionId === nextSelection.conditionId
-        && current.outcomeIndex === nextSelection.outcomeIndex
-      ) {
-        return current
-      }
-
-      return nextSelection
-    })
-  }, [])
-
-  return { querySelection, handleQuerySelectionChange }
-}
-
-function useActiveMarketView({
-  card,
-  marketViewCards,
-  initialMarketSlug,
-  initialMarketViewKey,
-}: {
-  card: SportsGamesCard
-  marketViewCards: SportsGamesCardMarketView[]
-  initialMarketSlug: string | null
-  initialMarketViewKey: SportsEventMarketViewKey | null
-}) {
-  const normalizedMarketViewCards = useMemo(
-    () => marketViewCards.length > 0
-      ? marketViewCards
-      : [{ key: 'gameLines' as const, label: 'Game Lines', card }],
-    [card, marketViewCards],
-  )
-  const initialMarketViewFromSlug = useMemo(
-    () => resolveMarketViewCardBySlug(normalizedMarketViewCards, initialMarketSlug)?.key ?? null,
-    [initialMarketSlug, normalizedMarketViewCards],
-  )
-  const resolvedInitialMarketViewKey = useMemo(() => {
-    if (
-      initialMarketViewFromSlug
-      && normalizedMarketViewCards.some(view => view.key === initialMarketViewFromSlug)
-    ) {
-      return initialMarketViewFromSlug
-    }
-
-    if (
-      initialMarketViewKey
-      && normalizedMarketViewCards.some(view => view.key === initialMarketViewKey)
-    ) {
-      return initialMarketViewKey
-    }
-
-    return normalizedMarketViewCards.find(view => view.key === 'gameLines')?.key
-      ?? normalizedMarketViewCards[0]?.key
-      ?? 'gameLines'
-  }, [initialMarketViewFromSlug, initialMarketViewKey, normalizedMarketViewCards])
-  const [activeMarketViewKey, setActiveMarketViewKey] = useState<SportsEventMarketViewKey>(resolvedInitialMarketViewKey)
-
-  useEffect(function resetActiveMarketViewWhenInitialChanges() {
-    setActiveMarketViewKey(resolvedInitialMarketViewKey)
-
-    return function noopResetActiveMarketViewCleanup() {}
-  }, [resolvedInitialMarketViewKey])
-
-  const activeMarketView = useMemo(
-    () => normalizedMarketViewCards.find(view => view.key === activeMarketViewKey)
-      ?? normalizedMarketViewCards.find(view => view.key === resolvedInitialMarketViewKey)
-      ?? normalizedMarketViewCards[0]
-      ?? null,
-    [activeMarketViewKey, normalizedMarketViewCards, resolvedInitialMarketViewKey],
-  )
-
-  return {
-    normalizedMarketViewCards,
-    activeMarketView,
-    activeMarketViewKey,
-    setActiveMarketViewKey,
-  }
-}
-
-function useEsportsSegmentTabState({
-  activeCard,
-  hasEsportsSegmentedLayout,
-  initialMarketSlug,
-}: {
-  activeCard: SportsGamesCard
-  hasEsportsSegmentedLayout: boolean
-  initialMarketSlug: string | null
-}) {
-  const esportsSegmentTabNumbers = useMemo(() => {
-    const numbers = new Set<number>()
-
-    activeCard.detailMarkets.forEach((market) => {
-      const mapNumber = parseEsportsSegmentNumber(market)
-      if (mapNumber != null) {
-        numbers.add(mapNumber)
-      }
-    })
-
-    return Array.from(numbers).sort((left, right) => left - right)
-  }, [activeCard.detailMarkets])
-  const initialEsportsSegmentTabKey = useMemo<EsportsLayoutTabKey>(() => {
-    if (!hasEsportsSegmentedLayout || !initialMarketSlug) {
-      return 'series'
-    }
-
-    const matchedMarket = activeCard.detailMarkets.find(market => market.slug === initialMarketSlug) ?? null
-    const mapNumber = parseEsportsSegmentNumber(matchedMarket)
-    if (mapNumber == null) {
-      return 'series'
-    }
-
-    return esportsSegmentTabNumbers.includes(mapNumber)
-      ? resolveEsportsSegmentTabKey(mapNumber)
-      : 'series'
-  }, [activeCard.detailMarkets, esportsSegmentTabNumbers, hasEsportsSegmentedLayout, initialMarketSlug])
-  const [activeEsportsSegmentTabKey, setActiveEsportsSegmentTabKey] = useState<EsportsLayoutTabKey>(initialEsportsSegmentTabKey)
-  const activeEsportsSegmentNumber = useMemo(
-    () => parseEsportsSegmentTabNumber(activeEsportsSegmentTabKey),
-    [activeEsportsSegmentTabKey],
-  )
-  const [activeSeriesPreviewSegmentNumber, setActiveSeriesPreviewSegmentNumber] = useState<number | null>(
-    esportsSegmentTabNumbers[0] ?? null,
-  )
-  const [activeSeriesSpreadPickerNumber, setActiveSeriesSpreadPickerNumber] = useState<number | null>(
-    esportsSegmentTabNumbers[0] ?? null,
-  )
-
-  useEffect(function resetActiveEsportsSegmentTabKey() {
-    setActiveEsportsSegmentTabKey(initialEsportsSegmentTabKey)
-
-    return function noopResetActiveEsportsSegmentTabKeyCleanup() {}
-  }, [initialEsportsSegmentTabKey])
-
-  useEffect(function clampActiveSeriesPreviewSegmentNumber() {
-    setActiveSeriesPreviewSegmentNumber(current => (
-      current != null && esportsSegmentTabNumbers.includes(current)
-        ? current
-        : (esportsSegmentTabNumbers[0] ?? null)
-    ))
-
-    return function noopClampActiveSeriesPreviewSegmentNumberCleanup() {}
-  }, [esportsSegmentTabNumbers])
-
-  useEffect(function clampActiveSeriesSpreadPickerNumber() {
-    setActiveSeriesSpreadPickerNumber(current => (
-      current != null && esportsSegmentTabNumbers.includes(current)
-        ? current
-        : (esportsSegmentTabNumbers[0] ?? null)
-    ))
-
-    return function noopClampActiveSeriesSpreadPickerNumberCleanup() {}
-  }, [esportsSegmentTabNumbers])
-
-  return {
-    esportsSegmentTabNumbers,
-    activeEsportsSegmentTabKey,
-    setActiveEsportsSegmentTabKey,
-    activeEsportsSegmentNumber,
-    activeSeriesPreviewSegmentNumber,
-    setActiveSeriesPreviewSegmentNumber,
-    activeSeriesSpreadPickerNumber,
-    setActiveSeriesSpreadPickerNumber,
-  }
-}
-
-function useUserPositionsQuery({ ownerAddress, activeCardId }: { ownerAddress: string | null, activeCardId: string }) {
-  return useQuery<UserPosition[]>({
-    queryKey: ['sports-event-user-positions', ownerAddress, activeCardId],
-    enabled: Boolean(ownerAddress),
-    staleTime: 1000 * 30,
-    gcTime: 1000 * 60 * 10,
-    refetchInterval: ownerAddress ? 15_000 : false,
-    refetchIntervalInBackground: true,
-    queryFn: ({ signal }) => fetchUserPositionsForMarket({
-      pageParam: 0,
-      userAddress: ownerAddress!,
-      status: 'active',
-      signal,
-    }),
-  })
-}
-
-function useRedeemModalState(activeCardId: string) {
-  const [claimedConditionIds, setClaimedConditionIds] = useState<Record<string, true>>({})
-  const [redeemSectionKey, setRedeemSectionKey] = useState<EventSectionKey | null>(null)
-  const [redeemDefaultConditionId, setRedeemDefaultConditionId] = useState<string | null>(null)
-
-  useEffect(function resetRedeemStateOnCardChange() {
-    setClaimedConditionIds(() => (activeCardId ? {} : {}))
-    setRedeemSectionKey(() => (activeCardId ? null : null))
-    setRedeemDefaultConditionId(() => (activeCardId ? null : null))
-
-    return function noopResetRedeemStateOnCardChangeCleanup() {}
-  }, [activeCardId])
-
-  return {
-    claimedConditionIds,
-    setClaimedConditionIds,
-    redeemSectionKey,
-    setRedeemSectionKey,
-    redeemDefaultConditionId,
-    setRedeemDefaultConditionId,
-  }
-}
-
-function useFormatButtonOdds(oddsFormat: OddsFormat) {
-  return useCallback((cents: number) => {
-    if (oddsFormat === 'price') {
-      return `${cents}¢`
-    }
-    return formatOddsFromCents(cents, oddsFormat)
-  }, [oddsFormat])
-}
-
-function useActiveCardPriceMap(activeCard: SportsGamesCard) {
-  const detailMarketByConditionId = useMemo(
-    () => new Map(activeCard.detailMarkets.map(market => [market.condition_id, market] as const)),
-    [activeCard.detailMarkets],
-  )
-  const activeCardButtonTokenIds = useMemo(() => {
-    const tokenIds = new Set<string>()
-
-    activeCard.buttons.forEach((button) => {
-      const market = detailMarketByConditionId.get(button.conditionId)
-      const outcome = market?.outcomes.find(currentOutcome => currentOutcome.outcome_index === button.outcomeIndex)
-        ?? market?.outcomes[button.outcomeIndex]
-
-      if (outcome?.token_id) {
-        tokenIds.add(String(outcome.token_id))
-      }
-    })
-
-    return Array.from(tokenIds)
-  }, [activeCard.buttons, detailMarketByConditionId])
-  const { data: buttonOrderBookSummaries } = useOrderBookSummaries(activeCardButtonTokenIds)
-  const buttonPriceCentsByKey = useMemo(() => {
-    const priceByKey = new Map<string, number>()
-
-    activeCard.buttons.forEach((button) => {
-      const market = detailMarketByConditionId.get(button.conditionId) ?? null
-      const outcome = market?.outcomes.find(currentOutcome => currentOutcome.outcome_index === button.outcomeIndex)
-        ?? market?.outcomes[button.outcomeIndex]
-      const cents = resolveOutcomePriceCents(
-        market,
-        button.outcomeIndex === OUTCOME_INDEX.NO ? OUTCOME_INDEX.NO : OUTCOME_INDEX.YES,
-        {
-          orderBookSummaries: buttonOrderBookSummaries,
-          side: ORDER_SIDE.BUY,
-        },
-      )
-      const selectionCents = resolveOutcomeSelectionPriceCents(market, outcome, {
-        orderBookSummaries: buttonOrderBookSummaries,
-        side: ORDER_SIDE.BUY,
-        fallbackIsNoOutcome: button.fallbackIsNoOutcome,
-      })
-      priceByKey.set(button.key, selectionCents ?? cents ?? button.cents)
-    })
-
-    return priceByKey
-  }, [activeCard.buttons, buttonOrderBookSummaries, detailMarketByConditionId])
-
-  return { detailMarketByConditionId, buttonPriceCentsByKey }
-}
-
-function useGroupedButtons({
-  activeCard,
-  detailMarketByConditionId,
-  hasEsportsSegmentedLayout,
-}: {
-  activeCard: SportsGamesCard
-  detailMarketByConditionId: Map<string, SportsGamesCard['detailMarkets'][number]>
-  hasEsportsSegmentedLayout: boolean
-}) {
-  const groupedButtons = useMemo(() => {
-    if (!hasEsportsSegmentedLayout) {
-      return groupButtonsByMarketType(activeCard.buttons)
-    }
-
-    const grouped: Record<SportsGamesMarketType, SportsGamesButton[]> = {
-      moneyline: [],
-      spread: [],
-      total: [],
-      btts: [],
-      binary: [],
-    }
-
-    activeCard.buttons.forEach((button) => {
-      const market = detailMarketByConditionId.get(button.conditionId)
-      if (!market) {
-        return
-      }
-
-      const sectionKey = resolveEventSectionKeyForButton(button, market)
-      if (!sectionKey) {
-        return
-      }
-
-      grouped[sectionKey].push(button)
-    })
-
-    return grouped
-  }, [activeCard.buttons, detailMarketByConditionId, hasEsportsSegmentedLayout])
-  const buttonByConditionAndOutcome = useMemo(() => {
-    const map = new Map<string, SportsGamesButton>()
-    activeCard.buttons.forEach((button) => {
-      map.set(`${button.conditionId}:${button.outcomeIndex}`, button)
-    })
-    return map
-  }, [activeCard.buttons])
-  const firstButtonByConditionId = useMemo(() => {
-    const map = new Map<string, SportsGamesButton>()
-    activeCard.buttons.forEach((button) => {
-      if (!map.has(button.conditionId)) {
-        map.set(button.conditionId, button)
-      }
-    })
-    return map
-  }, [activeCard.buttons])
-  const availableSections = useMemo(
-    () => SECTION_ORDER.filter(section => groupedButtons[section.key].length > 0),
-    [groupedButtons],
-  )
-  const sectionResolvedByKey = useMemo<Record<EventSectionKey, boolean>>(() => {
-    const resolved: Record<EventSectionKey, boolean> = {
-      moneyline: false,
-      spread: false,
-      total: false,
-      btts: false,
-    }
-
-    SECTION_ORDER.forEach((section) => {
-      const conditionIds = Array.from(new Set(groupedButtons[section.key].map(button => button.conditionId)))
-      if (conditionIds.length === 0) {
-        return
-      }
-
-      resolved[section.key] = conditionIds.every((conditionId) => {
-        const market = detailMarketByConditionId.get(conditionId)
-        return Boolean(market?.is_resolved || market?.condition?.resolved)
-      })
-    })
-
-    return resolved
-  }, [detailMarketByConditionId, groupedButtons])
-
-  return {
-    groupedButtons,
-    buttonByConditionAndOutcome,
-    firstButtonByConditionId,
-    availableSections,
-    sectionResolvedByKey,
-  }
-}
-
-function useClaimGroupsBySection({
-  activeCard,
-  buttonByConditionAndOutcome,
-  claimedConditionIds,
-  detailMarketByConditionId,
-  firstButtonByConditionId,
-  userPositions,
-}: {
-  activeCard: SportsGamesCard
-  buttonByConditionAndOutcome: Map<string, SportsGamesButton>
-  claimedConditionIds: Record<string, true>
-  detailMarketByConditionId: Map<string, SportsGamesCard['detailMarkets'][number]>
-  firstButtonByConditionId: Map<string, SportsGamesButton>
-  userPositions: UserPosition[] | undefined
-}) {
-  return useMemo<Record<EventSectionKey, SportsRedeemModalGroup[]>>(() => {
-    const bySection: Record<EventSectionKey, SportsRedeemModalGroup[]> = {
-      moneyline: [],
-      spread: [],
-      total: [],
-      btts: [],
-    }
-
-    if (!userPositions?.length) {
-      return bySection
-    }
-
-    const bySectionCondition = new Map<string, {
-      sectionKey: EventSectionKey
-      group: SportsRedeemModalGroup & { _indexSetCollection: Set<number> }
-    }>()
-
-    userPositions.forEach((position) => {
-      if (!position.redeemable) {
-        return
-      }
-
-      const conditionId = position.market?.condition_id
-      if (!conditionId || claimedConditionIds[conditionId]) {
-        return
-      }
-
-      const market = detailMarketByConditionId.get(conditionId)
-      const firstButton = firstButtonByConditionId.get(conditionId)
-      if (!market || !firstButton) {
-        return
-      }
-
-      const sectionKey = resolveEventSectionKeyForButton(firstButton, market)
-      if (!sectionKey) {
-        return
-      }
-
-      const shares = resolvePositionShares(position)
-      if (!(shares > 0)) {
-        return
-      }
-
-      const key = `${sectionKey}:${conditionId}`
-      let bucket = bySectionCondition.get(key)
-      if (!bucket) {
-        bucket = {
-          sectionKey,
-          group: {
-            conditionId,
-            title: resolveRedeemOptionLabel(activeCard, market, firstButton),
-            amount: 0,
-            indexSets: [],
-            isNegRisk: Boolean(market.neg_risk),
-            yesShares: 0,
-            noShares: 0,
-            positions: [],
-            _indexSetCollection: new Set<number>(),
-          },
-        }
-        bySectionCondition.set(key, bucket)
-      }
-      else if (market.neg_risk) {
-        bucket.group.isNegRisk = true
-      }
-
-      const outcomeIndex = resolveOutcomeIndexFromPosition(position)
-      const indexSet = resolveIndexSetFromOutcomeIndex(outcomeIndex)
-      if (indexSet) {
-        bucket.group._indexSetCollection.add(indexSet)
-      }
-
-      const positionButton = (outcomeIndex === OUTCOME_INDEX.YES || outcomeIndex === OUTCOME_INDEX.NO)
-        ? (buttonByConditionAndOutcome.get(`${conditionId}:${outcomeIndex}`) ?? firstButton)
-        : firstButton
-      const outcomeLabel = (outcomeIndex === OUTCOME_INDEX.YES || outcomeIndex === OUTCOME_INDEX.NO)
-        ? (market.outcomes.find(outcome => outcome.outcome_index === outcomeIndex)?.outcome_text
-          ?? position.outcome_text
-          ?? `Outcome ${outcomeIndex + 1}`)
-        : (position.outcome_text || 'Outcome')
-      const preferredButton = [positionButton, firstButton].find((button) => {
-        const normalizedLabel = button.label?.trim().toLowerCase()
-        return Boolean(normalizedLabel) && normalizedLabel !== 'yes' && normalizedLabel !== 'no'
-      })
-      const preferredButtonLabel = preferredButton
-        ? resolveRedeemOptionLabel(activeCard, market, preferredButton)
-        : null
-      const fallbackButtonLabel = [positionButton.label?.trim(), firstButton.label?.trim()].find((label) => {
-        const normalizedLabel = label?.toLowerCase()
-        return Boolean(label) && normalizedLabel !== 'yes' && normalizedLabel !== 'no'
-      })
-      const positionOptionLabel = preferredButtonLabel
-        || fallbackButtonLabel
-        || market.sports_group_item_title?.trim()
-        || market.short_title?.trim()
-        || market.title
-      const outcomeSideLabel = outcomeIndex === OUTCOME_INDEX.NO
-        ? 'No'
-        : outcomeIndex === OUTCOME_INDEX.YES
-          ? 'Yes'
-          : null
-      const positionLabel = outcomeSideLabel
-        ? `${positionOptionLabel || outcomeLabel} - ${outcomeSideLabel}`
-        : outcomeLabel
-      const tagAccent = resolveRedeemTagAccent(positionButton, outcomeIndex)
-
-      bucket.group.positions.push({
-        key: `${conditionId}-${outcomeLabel}-${bucket.group.positions.length}`,
-        label: positionLabel,
-        shares,
-        value: shares,
-        outcomeIndex,
-        badgeClassName: tagAccent.badgeClassName,
-        badgeStyle: tagAccent.badgeStyle,
-      })
-
-      if (bucket.group.isNegRisk) {
-        if (outcomeIndex === OUTCOME_INDEX.YES) {
-          bucket.group.yesShares = (bucket.group.yesShares ?? 0) + shares
-        }
-        else if (outcomeIndex === OUTCOME_INDEX.NO) {
-          bucket.group.noShares = (bucket.group.noShares ?? 0) + shares
-        }
-      }
-
-      bucket.group.amount += shares
-    })
-
-    bySectionCondition.forEach(({ sectionKey, group }) => {
-      if (group._indexSetCollection.size === 0) {
-        const market = detailMarketByConditionId.get(group.conditionId)
-        const winningOutcome = market?.outcomes.find(outcome => outcome.is_winning_outcome)
-        const fallbackIndexSet = resolveIndexSetFromOutcomeIndex(winningOutcome?.outcome_index)
-        if (fallbackIndexSet) {
-          group._indexSetCollection.add(fallbackIndexSet)
-        }
-      }
-
-      if (group._indexSetCollection.size === 0 || !(group.amount > 0)) {
-        return
-      }
-
-      bySection[sectionKey].push({
-        conditionId: group.conditionId,
-        title: group.title,
-        amount: group.amount,
-        indexSets: Array.from(group._indexSetCollection).sort((a, b) => a - b),
-        isNegRisk: group.isNegRisk,
-        yesShares: group.yesShares,
-        noShares: group.noShares,
-        positions: group.positions,
-      })
-    })
-
-    SECTION_ORDER.forEach((section) => {
-      bySection[section.key].sort((left, right) => right.amount - left.amount)
-    })
-
-    return bySection
-  }, [activeCard, buttonByConditionAndOutcome, claimedConditionIds, detailMarketByConditionId, firstButtonByConditionId, userPositions])
-}
-
-function useMarketSlugToButtonKey({
-  activeCard,
-  querySelection,
-  initialMarketSlug,
-}: {
-  activeCard: SportsGamesCard
-  querySelection: SportsEventQuerySelection
-  initialMarketSlug: string | null
-}) {
-  return useMemo(() => {
-    const requestedConditionId = querySelection.conditionId
-    const requestedOutcomeIndex = querySelection.outcomeIndex
-
-    function resolveButtonKeyForConditionId(conditionId: string) {
-      if (requestedOutcomeIndex !== null) {
-        const exactMatch = activeCard.buttons.find(button =>
-          button.conditionId === conditionId && button.outcomeIndex === requestedOutcomeIndex,
-        )
-        if (exactMatch) {
-          return exactMatch.key
-        }
-      }
-
-      return activeCard.buttons.find(button => button.conditionId === conditionId)?.key ?? null
-    }
-
-    if (requestedConditionId) {
-      return resolveButtonKeyForConditionId(requestedConditionId)
-    }
-
-    if (!initialMarketSlug) {
-      return null
-    }
-
-    const matchedMarket = activeCard.detailMarkets.find(market => market.slug === initialMarketSlug)
-    if (!matchedMarket) {
-      return null
-    }
-
-    return resolveButtonKeyForConditionId(matchedMarket.condition_id)
-  }, [
-    activeCard.buttons,
-    activeCard.detailMarkets,
-    initialMarketSlug,
-    querySelection.conditionId,
-    querySelection.outcomeIndex,
-  ])
-}
-
-function useAuxiliaryMarketCards({
-  activeCard,
-  baseUsesSectionLayout,
-  hasEsportsSegmentedLayout,
-  activeEsportsSegmentTabKey,
-  activeEsportsSegmentNumber,
-  activeSeriesPreviewSegmentNumber,
-}: {
-  activeCard: SportsGamesCard
-  baseUsesSectionLayout: boolean
-  hasEsportsSegmentedLayout: boolean
-  activeEsportsSegmentTabKey: EsportsLayoutTabKey
-  activeEsportsSegmentNumber: number | null
-  activeSeriesPreviewSegmentNumber: number | null
-}) {
-  const auxiliaryMarketCards = useMemo<AuxiliaryMarketPanel[]>(() => {
-    const buttonsByConditionId = new Map<string, SportsGamesButton[]>()
-
-    activeCard.buttons.forEach((button) => {
-      const currentButtons = buttonsByConditionId.get(button.conditionId) ?? []
-      currentButtons.push(button)
-      buttonsByConditionId.set(button.conditionId, currentButtons)
-    })
-
-    const panelsByKey = new Map<string, AuxiliaryMarketPanel>()
-
-    activeCard.detailMarkets.forEach((market) => {
-      const buttons = sortAuxiliaryButtons(buttonsByConditionId.get(market.condition_id) ?? [])
-
-      if (buttons.length === 0) {
-        return
-      }
-
-      const isSegmentedEsportsMarket = hasEsportsSegmentedLayout && parseEsportsSegmentNumber(market) != null
-
-      if (baseUsesSectionLayout && buttons[0]?.marketType !== 'binary' && !isSegmentedEsportsMarket) {
-        return
-      }
-
-      const mapNumber = hasEsportsSegmentedLayout && isSegmentedEsportsMarket
-        ? parseEsportsSegmentNumber(market)
-        : null
-      const panelKey = mapNumber != null
-        ? `${activeCard.id}:${normalizeSportsMarketType(market.sports_market_type)}:map-${mapNumber}`
-        : resolveSportsAuxiliaryMarketGroupKey(market)
-      const existingPanel = panelsByKey.get(panelKey)
-      if (existingPanel) {
-        existingPanel.markets.push(market)
-        existingPanel.buttons.push(...buttons)
-        existingPanel.volume += Number(market.volume ?? 0)
-        return
-      }
-
-      panelsByKey.set(panelKey, {
-        key: panelKey,
-        title: '',
-        markets: [market],
-        buttons: [...buttons],
-        volume: Number(market.volume ?? 0),
-        mapNumber,
-      })
-    })
-
-    return Array.from(panelsByKey.values())
-      .map(panel => ({
-        ...panel,
-        title: panel.mapNumber != null
-          ? resolveEsportsSegmentPanelTitle(panel.markets)
-          : resolveSportsAuxiliaryMarketTitle(panel.markets),
-        buttons: sortAuxiliaryButtons(dedupeAuxiliaryButtons(panel.buttons)),
-      }))
-      .sort((left, right) => {
-        const mapComparison = (left.mapNumber ?? 0) - (right.mapNumber ?? 0)
-        if (mapComparison !== 0) {
-          return mapComparison
-        }
-
-        const segmentTypeComparison = resolveEsportsSegmentPanelSortOrder(left.markets)
-          - resolveEsportsSegmentPanelSortOrder(right.markets)
-        if (segmentTypeComparison !== 0) {
-          return segmentTypeComparison
-        }
-
-        const thresholdComparison = resolveAuxiliaryPanelThreshold(left.markets)
-          - resolveAuxiliaryPanelThreshold(right.markets)
-        if (thresholdComparison !== 0) {
-          return thresholdComparison
-        }
-
-        const timestampComparison = resolveAuxiliaryPanelCreatedAt(left.markets) - resolveAuxiliaryPanelCreatedAt(right.markets)
-        if (timestampComparison !== 0) {
-          return timestampComparison
-        }
-
-        return left.title.localeCompare(right.title)
-      })
-  }, [activeCard.buttons, activeCard.detailMarkets, activeCard.id, baseUsesSectionLayout, hasEsportsSegmentedLayout])
-  const renderedAuxiliaryMarketCards = useMemo(() => {
-    if (!hasEsportsSegmentedLayout) {
-      return auxiliaryMarketCards
-    }
-
-    if (activeEsportsSegmentTabKey === 'series') {
-      return auxiliaryMarketCards.filter(entry => entry.mapNumber == null)
-    }
-
-    if (activeEsportsSegmentNumber == null) {
-      return []
-    }
-
-    return auxiliaryMarketCards.filter(entry => entry.mapNumber === activeEsportsSegmentNumber)
-  }, [activeEsportsSegmentNumber, activeEsportsSegmentTabKey, auxiliaryMarketCards, hasEsportsSegmentedLayout])
-  const seriesPreviewSegmentWinnerPanels = useMemo(() => {
-    if (!hasEsportsSegmentedLayout) {
-      return [] as AuxiliaryMarketPanel[]
-    }
-
-    return auxiliaryMarketCards.filter(entry =>
-      entry.mapNumber != null
-      && entry.markets.some(market => isSegmentedEsportsChildMoneylineMarket(market))
-      && entry.buttons.every(button => button.marketType === 'moneyline'),
-    )
-  }, [auxiliaryMarketCards, hasEsportsSegmentedLayout])
-  const activeSeriesPreviewSegmentWinnerPanel = useMemo(() => {
-    if (seriesPreviewSegmentWinnerPanels.length === 0) {
-      return null
-    }
-
-    return seriesPreviewSegmentWinnerPanels.find(entry => entry.mapNumber === activeSeriesPreviewSegmentNumber)
-      ?? seriesPreviewSegmentWinnerPanels[0]
-      ?? null
-  }, [activeSeriesPreviewSegmentNumber, seriesPreviewSegmentWinnerPanels])
-  const auxiliaryPanelKeyByButtonKey = useMemo(() => {
-    const map = new Map<string, string>()
-
-    auxiliaryMarketCards.forEach((entry) => {
-      entry.buttons.forEach((button) => {
-        map.set(button.key, entry.key)
-      })
-    })
-
-    return map
-  }, [auxiliaryMarketCards])
-  const seriesWinnerSegmentPickerOptions = useMemo(
-    () => seriesPreviewSegmentWinnerPanels
-      .filter((panel): panel is AuxiliaryMarketPanel & { mapNumber: number } => panel.mapNumber != null)
-      .map(panel => ({
-        key: `winner-segment-${panel.mapNumber}`,
-        label: `${panel.mapNumber}`,
-        number: panel.mapNumber,
-      })),
-    [seriesPreviewSegmentWinnerPanels],
-  )
-
-  return {
-    auxiliaryMarketCards,
-    renderedAuxiliaryMarketCards,
-    seriesPreviewSegmentWinnerPanels,
-    activeSeriesPreviewSegmentWinnerPanel,
-    auxiliaryPanelKeyByButtonKey,
-    seriesWinnerSegmentPickerOptions,
-  }
-}
-
-function useSelectionState({
-  activeCard,
-  auxiliaryPanelsForSelection,
-  detailMarketByConditionId,
-  groupedButtons,
-  usesSectionLayout,
-  marketSlugToButtonKey,
-  renderedAuxiliaryMarketCards,
-}: {
-  activeCard: SportsGamesCard
-  auxiliaryPanelsForSelection: AuxiliaryMarketPanel[]
-  detailMarketByConditionId: Map<string, SportsGamesCard['detailMarkets'][number]>
-  groupedButtons: Record<SportsGamesMarketType, SportsGamesButton[]>
-  usesSectionLayout: boolean
-  marketSlugToButtonKey: string | null
-  renderedAuxiliaryMarketCards: AuxiliaryMarketPanel[]
-}) {
-  const [selectedButtonBySection, setSelectedButtonBySection] = useState<Record<EventSectionKey, string | null>>({
-    moneyline: null,
-    spread: null,
-    total: null,
-    btts: null,
-  })
-  const [selectedAuxiliaryButtonByConditionId, setSelectedAuxiliaryButtonByConditionId] = useState<
-    Record<string, string | null>
-  >({})
-  const [activeTradeButtonKey, setActiveTradeButtonKey] = useState<string | null>(null)
-  const [openSectionKey, setOpenSectionKey] = useState<EventSectionKey | null>(null)
-  const [openAuxiliaryConditionId, setOpenAuxiliaryConditionId] = useState<string | null>(null)
-  const [tabBySection, setTabBySection] = useState<Record<EventSectionKey, DetailsTab>>({
-    moneyline: 'orderBook',
-    spread: 'orderBook',
-    total: 'orderBook',
-    btts: 'orderBook',
-  })
-  const [tabByAuxiliaryConditionId, setTabByAuxiliaryConditionId] = useState<Record<string, DetailsTab>>({})
-  const previousCardIdRef = useRef<string | null>(null)
-  const appliedMarketSlugSelectionRef = useRef<string | null>(null)
-
-  useEffect(function syncSelectionsOnActiveCardChange() {
-    const isNewCard = previousCardIdRef.current !== activeCard.id
-    previousCardIdRef.current = activeCard.id
-    const marketSlugSelectionSignature = buildMarketSlugSelectionSignature({
-      activeCardId: activeCard.id,
-      marketSlugToButtonKey,
-      usesSectionLayout,
-    })
-    const shouldApplyMarketSlugSelection = marketSlugSelectionSignature !== null
-      && appliedMarketSlugSelectionRef.current !== marketSlugSelectionSignature
-
-    if (!marketSlugSelectionSignature) {
-      appliedMarketSlugSelectionRef.current = null
-    }
-
-    const defaultSelectedByCondition = auxiliaryPanelsForSelection.reduce<Record<string, string | null>>((acc, entry) => {
-      const marketMatchedButton = shouldApplyMarketSlugSelection
-        && marketSlugToButtonKey
-        && entry.buttons.some(button => button.key === marketSlugToButtonKey)
-        ? marketSlugToButtonKey
-        : null
-      const defaultButtonKey = entry.buttons[0]?.key ?? null
-      acc[entry.key] = marketMatchedButton ?? defaultButtonKey
-      return acc
-    }, {})
-
-    setSelectedAuxiliaryButtonByConditionId((current) => {
-      if (isNewCard) {
-        return areRecordValuesEqual(current, defaultSelectedByCondition)
-          ? current
-          : defaultSelectedByCondition
-      }
-
-      const next = { ...defaultSelectedByCondition }
-      Object.entries(current).forEach(([conditionId, buttonKey]) => {
-        if (!buttonKey) {
-          return
-        }
-
-        const matchedEntry = auxiliaryPanelsForSelection.find(entry => entry.key === conditionId)
-        if (!matchedEntry) {
-          return
-        }
-
-        if (matchedEntry.buttons.some(button => button.key === buttonKey)) {
-          next[conditionId] = buttonKey
-        }
-      })
-
-      if (shouldApplyMarketSlugSelection && marketSlugToButtonKey) {
-        const matchedEntry = auxiliaryPanelsForSelection.find(entry =>
-          entry.buttons.some(button => button.key === marketSlugToButtonKey),
-        )
-        if (matchedEntry) {
-          next[matchedEntry.key] = marketSlugToButtonKey
-        }
-      }
-
-      return areRecordValuesEqual(current, next) ? current : next
-    })
-
-    setTabByAuxiliaryConditionId((current) => {
-      const next = { ...current }
-      let changed = false
-      auxiliaryPanelsForSelection.forEach(({ key }) => {
-        if (!next[key]) {
-          next[key] = 'orderBook'
-          changed = true
-        }
-      })
-      return changed ? next : current
-    })
-
-    const marketMatchedAuxiliaryConditionId = shouldApplyMarketSlugSelection && marketSlugToButtonKey
-      ? auxiliaryPanelsForSelection.find(entry => entry.buttons.some(button => button.key === marketSlugToButtonKey))?.key ?? null
-      : null
-
-    if (!usesSectionLayout) {
-      const defaultTradeButton = (shouldApplyMarketSlugSelection ? marketSlugToButtonKey : null)
-        ?? renderedAuxiliaryMarketCards[0]?.buttons[0]?.key
-        ?? auxiliaryPanelsForSelection[0]?.buttons[0]?.key
-        ?? resolveDefaultConditionId(activeCard)
-
-      setActiveTradeButtonKey((current) => {
-        if (
-          shouldApplyMarketSlugSelection
-          && marketSlugToButtonKey
-          && activeCard.buttons.some(button => button.key === marketSlugToButtonKey)
-        ) {
-          return marketSlugToButtonKey
-        }
-
-        if (!isNewCard && current && activeCard.buttons.some(button => button.key === current)) {
-          return current
-        }
-
-        return defaultTradeButton
-      })
-
-      setOpenSectionKey(() => (activeCard.id ? null : null))
-      setOpenAuxiliaryConditionId((current) => {
-        if (marketMatchedAuxiliaryConditionId) {
-          return marketMatchedAuxiliaryConditionId
-        }
-
-        if (isNewCard) {
-          return null
-        }
-
-        if (current && renderedAuxiliaryMarketCards.some(entry => entry.key === current)) {
-          return current
-        }
-
-        return null
-      })
-      if (marketSlugSelectionSignature) {
-        appliedMarketSlugSelectionRef.current = marketSlugSelectionSignature
-      }
-      return
-    }
-
-    const defaultSelectedBySection: Record<EventSectionKey, string | null> = {
-      moneyline: null,
-      spread: null,
-      total: null,
-      btts: null,
-    }
-
-    for (const section of SECTION_ORDER) {
-      const firstButton = groupedButtons[section.key][0] ?? null
-      defaultSelectedBySection[section.key] = firstButton?.key ?? null
-    }
-
-    if (shouldApplyMarketSlugSelection && marketSlugToButtonKey) {
-      const marketButton = activeCard.buttons.find(button => button.key === marketSlugToButtonKey)
-      const market = marketButton
-        ? (detailMarketByConditionId.get(marketButton.conditionId) ?? null)
-        : null
-      const sectionKey = resolveEventSectionKeyForButton(marketButton, market)
-      if (marketButton && sectionKey) {
-        defaultSelectedBySection[sectionKey] = marketButton.key
-      }
-    }
-
-    setSelectedButtonBySection((current) => {
-      if (isNewCard) {
-        return areRecordValuesEqual(current, defaultSelectedBySection)
-          ? current
-          : defaultSelectedBySection
-      }
-
-      const next: Record<EventSectionKey, string | null> = {
-        ...defaultSelectedBySection,
-      }
-
-      for (const section of SECTION_ORDER) {
-        const currentButtonKey = current[section.key]
-        if (!currentButtonKey) {
-          continue
-        }
-
-        const stillExists = groupedButtons[section.key].some(button => button.key === currentButtonKey)
-        if (stillExists) {
-          next[section.key] = currentButtonKey
-        }
-      }
-
-      if (shouldApplyMarketSlugSelection && marketSlugToButtonKey) {
-        const marketButton = activeCard.buttons.find(button => button.key === marketSlugToButtonKey)
-        const market = marketButton
-          ? (detailMarketByConditionId.get(marketButton.conditionId) ?? null)
-          : null
-        const sectionKey = resolveEventSectionKeyForButton(marketButton, market)
-        if (marketButton && sectionKey) {
-          next[sectionKey] = marketButton.key
-        }
-      }
-
-      return areRecordValuesEqual(current, next) ? current : next
-    })
-
-    const defaultTradeButton = (shouldApplyMarketSlugSelection ? marketSlugToButtonKey : null)
-      ?? defaultSelectedBySection.moneyline
-      ?? defaultSelectedBySection.spread
-      ?? defaultSelectedBySection.total
-      ?? defaultSelectedBySection.btts
-      ?? resolveDefaultConditionId(activeCard)
-
-    setActiveTradeButtonKey((current) => {
-      if (shouldApplyMarketSlugSelection && marketSlugToButtonKey) {
-        const matchesMarketSlug = activeCard.buttons.some(button => button.key === marketSlugToButtonKey)
-        if (matchesMarketSlug) {
-          return marketSlugToButtonKey
-        }
-      }
-
-      if (!isNewCard && current) {
-        const stillExists = activeCard.buttons.some(button => button.key === current)
-        if (stillExists) {
-          return current
-        }
-      }
-
-      return defaultTradeButton
-    })
-
-    setOpenSectionKey((current) => {
-      if (isNewCard) {
-        return null
-      }
-      if (current && groupedButtons[current].length > 0) {
-        return current
-      }
-      return null
-    })
-    setOpenAuxiliaryConditionId((current) => {
-      if (marketMatchedAuxiliaryConditionId) {
-        return marketMatchedAuxiliaryConditionId
-      }
-
-      if (!isNewCard && current && renderedAuxiliaryMarketCards.some(entry => entry.key === current)) {
-        return current
-      }
-
-      return null
-    })
-    if (marketSlugSelectionSignature) {
-      appliedMarketSlugSelectionRef.current = marketSlugSelectionSignature
-    }
-
-    return function noopSyncSelectionsOnActiveCardChangeCleanup() {}
-  }, [
-    activeCard,
-    activeCard.id,
-    activeCard.buttons,
-    auxiliaryPanelsForSelection,
-    detailMarketByConditionId,
-    groupedButtons,
-    usesSectionLayout,
-    marketSlugToButtonKey,
-    renderedAuxiliaryMarketCards,
-  ])
-
-  return {
-    selectedButtonBySection,
-    setSelectedButtonBySection,
-    selectedAuxiliaryButtonByConditionId,
-    setSelectedAuxiliaryButtonByConditionId,
-    activeTradeButtonKey,
-    setActiveTradeButtonKey,
-    openSectionKey,
-    setOpenSectionKey,
-    openAuxiliaryConditionId,
-    setOpenAuxiliaryConditionId,
-    tabBySection,
-    setTabBySection,
-    tabByAuxiliaryConditionId,
-    setTabByAuxiliaryConditionId,
-  }
-}
-
-function useDerivedActiveCard({
-  card,
-  activeMarketView,
-  vertical,
-}: {
-  card: SportsGamesCard
-  activeMarketView: SportsGamesCardMarketView | null
-  vertical: SportsVertical
-}) {
-  const activeSourceCard = activeMarketView?.card ?? card
-  const activeCard = useMemo(
-    () => resolveNormalizedSegmentedEsportsCard(activeSourceCard, vertical),
-    [activeSourceCard, vertical],
-  )
-  const heroGroupedButtons = useMemo(() => groupButtonsByMarketType(card.buttons), [card.buttons])
-  const { singular: segmentLabel, plural: segmentPluralLabel } = useMemo(
-    () => resolveEsportsSegmentLabels(activeCard),
-    [activeCard],
-  )
-  const isGameLinesView = (activeMarketView?.key ?? 'gameLines') === 'gameLines'
-  const isHalftimeResultView = activeMarketView?.key === 'halftimeResult'
-  const baseUsesSectionLayout = isGameLinesView || isHalftimeResultView
-  const hasEsportsSegmentedLayout = useMemo(
-    () => baseUsesSectionLayout
-      && isSegmentedEsportsEventCard(activeCard, vertical)
-      && activeCard.detailMarkets.some(market => parseEsportsSegmentNumber(market) != null),
-    [activeCard, baseUsesSectionLayout, vertical],
-  )
-
-  return {
-    activeCard,
-    heroGroupedButtons,
-    segmentLabel,
-    segmentPluralLabel,
-    isHalftimeResultView,
-    baseUsesSectionLayout,
-    hasEsportsSegmentedLayout,
-  }
-}
-
-function useSelectedSectionButtons({
-  activeCard,
-  groupedButtons,
-  selectedButtonBySection,
-}: {
-  activeCard: SportsGamesCard
-  groupedButtons: Record<SportsGamesMarketType, SportsGamesButton[]>
-  selectedButtonBySection: Record<EventSectionKey, string | null>
-}) {
-  const selectedSpreadSectionButton = useMemo(() => {
-    if (selectedButtonBySection.spread) {
-      const selected = activeCard.buttons.find(button => button.key === selectedButtonBySection.spread) ?? null
-      if (selected?.marketType === 'spread') {
-        return selected
-      }
-    }
-
-    return groupedButtons.spread[0] ?? null
-  }, [activeCard.buttons, groupedButtons.spread, selectedButtonBySection.spread])
-  const selectedTotalSectionButton = useMemo(() => {
-    if (selectedButtonBySection.total) {
-      const selected = activeCard.buttons.find(button => button.key === selectedButtonBySection.total) ?? null
-      if (selected?.marketType === 'total') {
-        return selected
-      }
-    }
-
-    return groupedButtons.total[0] ?? null
-  }, [activeCard.buttons, groupedButtons.total, selectedButtonBySection.total])
-
-  return { selectedSpreadSectionButton, selectedTotalSectionButton }
-}
-
-function useSectionActions({
-  isMobile,
-  setIsMobileOrderPanelOpen,
-  setSelectedButtonBySection,
-  setActiveTradeButtonKey,
-  setOpenSectionKey,
-  setOpenAuxiliaryConditionId,
-  setActiveSeriesPreviewSegmentNumber,
-  setActiveSeriesSpreadPickerNumber,
-  seriesSpreadSegmentOptions,
-  seriesTotalLinePickerOptions,
-  selectedSpreadSectionButton,
-  selectedTotalSectionButton,
-  activeSeriesSpreadSegmentOption,
-  activeSeriesSpreadConditionId,
-  activeSeriesTotalLineOption,
-}: {
-  isMobile: boolean
-  setIsMobileOrderPanelOpen: (open: boolean) => void
-  setSelectedButtonBySection: React.Dispatch<React.SetStateAction<Record<EventSectionKey, string | null>>>
-  setActiveTradeButtonKey: React.Dispatch<React.SetStateAction<string | null>>
-  setOpenSectionKey: React.Dispatch<React.SetStateAction<EventSectionKey | null>>
-  setOpenAuxiliaryConditionId: React.Dispatch<React.SetStateAction<string | null>>
-  setActiveSeriesPreviewSegmentNumber: React.Dispatch<React.SetStateAction<number | null>>
-  setActiveSeriesSpreadPickerNumber: React.Dispatch<React.SetStateAction<number | null>>
-  seriesSpreadSegmentOptions: Array<{
-    number: number
-    conditionIds: string[]
-    buttonsByConditionId: Map<string, SportsGamesButton[]>
-  }>
-  seriesTotalLinePickerOptions: SportsLinePickerOption[]
-  selectedSpreadSectionButton: SportsGamesButton | null
-  selectedTotalSectionButton: SportsGamesButton | null
-  activeSeriesSpreadSegmentOption: {
-    number: number
-    conditionIds: string[]
-    buttonsByConditionId: Map<string, SportsGamesButton[]>
-  } | null
-  activeSeriesSpreadConditionId: string | null
-  activeSeriesTotalLineOption: SportsLinePickerOption | null
-}) {
-  const updateSectionSelection = useCallback((
-    sectionKey: EventSectionKey,
-    buttonKey: string,
-    options?: { panelMode?: 'full' | 'partial' | 'preserve' },
-  ) => {
-    setSelectedButtonBySection((current) => {
-      if (current[sectionKey] === buttonKey) {
-        return current
-      }
-      return {
-        ...current,
-        [sectionKey]: buttonKey,
-      }
-    })
-
-    setActiveTradeButtonKey(buttonKey)
-
-    const panelMode = options?.panelMode ?? 'full'
-    const shouldOpenMobileSheetOnly = isMobile && panelMode === 'full'
-
-    if (shouldOpenMobileSheetOnly) {
-      setIsMobileOrderPanelOpen(true)
-    }
-
-    if (panelMode === 'full' && !shouldOpenMobileSheetOnly) {
-      setOpenAuxiliaryConditionId(null)
-      setOpenSectionKey(sectionKey)
-    }
-  }, [
-    isMobile,
-    setIsMobileOrderPanelOpen,
-    setActiveTradeButtonKey,
-    setOpenAuxiliaryConditionId,
-    setOpenSectionKey,
-    setSelectedButtonBySection,
-  ])
-
-  const handlePickSeriesPreviewSegmentNumber = useCallback((number: number) => {
-    setActiveSeriesPreviewSegmentNumber(number)
-  }, [setActiveSeriesPreviewSegmentNumber])
-
-  const handlePickSeriesSpreadSegmentNumber = useCallback((number: number) => {
-    setActiveSeriesSpreadPickerNumber(number)
-
-    const spreadOption = seriesSpreadSegmentOptions.find(option => option.number === number)
-    if (!spreadOption) {
-      return
-    }
-
-    const currentSpreadConditionId = selectedSpreadSectionButton?.conditionId ?? null
-    const preferredConditionId = currentSpreadConditionId && spreadOption.buttonsByConditionId.has(currentSpreadConditionId)
-      ? currentSpreadConditionId
-      : spreadOption.conditionIds[0] ?? null
-    if (!preferredConditionId) {
-      return
-    }
-
-    const buttons = spreadOption.buttonsByConditionId.get(preferredConditionId) ?? []
-    const preferredButton = resolvePreferredLinePickerButton(buttons, selectedSpreadSectionButton)
-    if (!preferredButton) {
-      return
-    }
-
-    updateSectionSelection('spread', preferredButton.key, { panelMode: 'preserve' })
-  }, [selectedSpreadSectionButton, seriesSpreadSegmentOptions, setActiveSeriesSpreadPickerNumber, updateSectionSelection])
-  const handlePickSeriesTotalLineValue = useCallback((lineValue: number) => {
-    const option = seriesTotalLinePickerOptions.find(candidate => candidate.lineValue === lineValue) ?? null
-    if (!option) {
-      return
-    }
-
-    const preferredButton = resolvePreferredLinePickerButton(option.buttons, selectedTotalSectionButton)
-    if (!preferredButton) {
-      return
-    }
-
-    updateSectionSelection('total', preferredButton.key, { panelMode: 'preserve' })
-  }, [selectedTotalSectionButton, seriesTotalLinePickerOptions, updateSectionSelection])
-
-  const resolveSeriesSpreadSelectedButtonKey = useCallback(() => {
-    if (!activeSeriesSpreadSegmentOption) {
-      return null
-    }
-
-    const preferredConditionId = activeSeriesSpreadConditionId
-      ?? activeSeriesSpreadSegmentOption.conditionIds[0]
-      ?? null
-    if (!preferredConditionId) {
-      return null
-    }
-
-    const buttons = activeSeriesSpreadSegmentOption.buttonsByConditionId.get(preferredConditionId) ?? []
-
-    return resolvePreferredLinePickerButton(buttons, selectedSpreadSectionButton)?.key
-      ?? null
-  }, [activeSeriesSpreadConditionId, activeSeriesSpreadSegmentOption, selectedSpreadSectionButton])
-  const resolveSeriesTotalSelectedButtonKey = useCallback(() => {
-    if (!activeSeriesTotalLineOption) {
-      return null
-    }
-
-    return resolvePreferredLinePickerButton(activeSeriesTotalLineOption.buttons, selectedTotalSectionButton)?.key
-      ?? null
-  }, [activeSeriesTotalLineOption, selectedTotalSectionButton])
-
-  return {
-    updateSectionSelection,
-    handlePickSeriesPreviewSegmentNumber,
-    handlePickSeriesSpreadSegmentNumber,
-    handlePickSeriesTotalLineValue,
-    resolveSeriesSpreadSelectedButtonKey,
-    resolveSeriesTotalSelectedButtonKey,
-  }
-}
-
-function useSectionDerivedData({
-  activeCard,
-  groupedButtons,
-  claimGroupsBySection,
-  auxiliaryPanelsForSelection,
-  redeemSectionKey,
-  setRedeemDefaultConditionId,
-  setRedeemSectionKey,
-}: {
-  activeCard: SportsGamesCard
-  groupedButtons: Record<SportsGamesMarketType, SportsGamesButton[]>
-  claimGroupsBySection: Record<EventSectionKey, SportsRedeemModalGroup[]>
-  auxiliaryPanelsForSelection: AuxiliaryMarketPanel[]
-  redeemSectionKey: EventSectionKey | null
-  setRedeemDefaultConditionId: React.Dispatch<React.SetStateAction<string | null>>
-  setRedeemSectionKey: React.Dispatch<React.SetStateAction<EventSectionKey | null>>
-}) {
-  const sectionVolumes = useMemo(() => {
-    const byConditionId = new Map(activeCard.detailMarkets.map(market => [market.condition_id, market] as const))
-    const volumes: Record<EventSectionKey, number> = {
-      moneyline: 0,
-      spread: 0,
-      total: 0,
-      btts: 0,
-    }
-
-    for (const section of SECTION_ORDER) {
-      const conditionIds = Array.from(new Set(groupedButtons[section.key].map(button => button.conditionId)))
-      volumes[section.key] = conditionIds.reduce((sum, conditionId) => {
-        const market = byConditionId.get(conditionId)
-        return sum + (Number(market?.volume ?? 0) || 0)
-      }, 0)
-    }
-
-    return volumes
-  }, [activeCard.detailMarkets, groupedButtons])
-
-  const sectionConditionIdsByKey = useMemo<Record<EventSectionKey, Set<string>>>(() => {
-    return {
-      moneyline: new Set(groupedButtons.moneyline.map(button => button.conditionId)),
-      spread: new Set(groupedButtons.spread.map(button => button.conditionId)),
-      total: new Set(groupedButtons.total.map(button => button.conditionId)),
-      btts: new Set(groupedButtons.btts.map(button => button.conditionId)),
-    }
-  }, [groupedButtons])
-
-  const allCardConditionIds = useMemo(
-    () => new Set(activeCard.detailMarkets.map(market => market.condition_id)),
-    [activeCard.detailMarkets],
-  )
-  const redeemSectionConfig = useMemo(
-    () => (redeemSectionKey ? SECTION_ORDER.find(section => section.key === redeemSectionKey) ?? null : null),
-    [redeemSectionKey],
-  )
-  const redeemModalSections = useMemo<SportsRedeemModalSection[]>(
-    () =>
-      SECTION_ORDER
-        .map(section => ({
-          key: section.key,
-          label: section.label,
-          groups: claimGroupsBySection[section.key],
-        }))
-        .filter(section => section.groups.length > 0),
-    [claimGroupsBySection],
-  )
-  const auxiliaryResolvedByConditionId = useMemo(
-    () => new Map(auxiliaryPanelsForSelection.map(entry => [
-      entry.key,
-      entry.markets.every(market => Boolean(market.is_resolved || market.condition?.resolved)),
-    ] as const)),
-    [auxiliaryPanelsForSelection],
-  )
-  const auxiliaryClaimGroupsByConditionId = useMemo(
-    () => new Map(claimGroupsBySection.moneyline.map(group => [group.conditionId, group] as const)),
-    [claimGroupsBySection],
-  )
-  const handleOpenRedeemForCondition = useCallback((conditionId: string) => {
-    const normalizedConditionId = conditionId.trim()
-    if (!normalizedConditionId) {
-      return
-    }
-
-    const matchedSection = SECTION_ORDER.find(section =>
-      claimGroupsBySection[section.key].some(group => group.conditionId === normalizedConditionId),
-    ) ?? SECTION_ORDER.find(section => sectionConditionIdsByKey[section.key].has(normalizedConditionId))
-    ?? SECTION_ORDER.find(section => claimGroupsBySection[section.key].length > 0)
-    ?? null
-
-    if (!matchedSection) {
-      return
-    }
-
-    setRedeemDefaultConditionId(normalizedConditionId)
-    setRedeemSectionKey(matchedSection.key)
-  }, [claimGroupsBySection, sectionConditionIdsByKey, setRedeemDefaultConditionId, setRedeemSectionKey])
-
-  return {
-    sectionVolumes,
-    sectionConditionIdsByKey,
-    allCardConditionIds,
-    redeemSectionConfig,
-    redeemModalSections,
-    auxiliaryResolvedByConditionId,
-    auxiliaryClaimGroupsByConditionId,
-    handleOpenRedeemForCondition,
-  }
-}
-
-function useActiveTradeContext({
-  activeCard,
-  usesSectionLayout,
-  activeTradeButtonKey,
-  openSectionKey,
-  openAuxiliaryConditionId,
-  selectedButtonBySection,
-  selectedAuxiliaryButtonByConditionId,
-  marketSlugToButtonKey,
-  renderedAuxiliaryMarketCards,
-  moneylineButtonKey,
-  orderEventId,
-  orderMarketConditionId,
-  orderOutcomeIndex,
-}: {
-  activeCard: SportsGamesCard
-  usesSectionLayout: boolean
-  activeTradeButtonKey: string | null
-  openSectionKey: EventSectionKey | null
-  openAuxiliaryConditionId: string | null
-  selectedButtonBySection: Record<EventSectionKey, string | null>
-  selectedAuxiliaryButtonByConditionId: Record<string, string | null>
-  marketSlugToButtonKey: string | null
-  renderedAuxiliaryMarketCards: AuxiliaryMarketPanel[]
-  moneylineButtonKey: string | null
-  orderEventId: string | null
-  orderMarketConditionId: string | null
-  orderOutcomeIndex: number | null
-}) {
-  const orderSelectionSyncKey = useMemo(() => {
-    if (!orderEventId || !orderMarketConditionId || orderOutcomeIndex == null) {
-      return null
-    }
-
-    return `${orderEventId}:${orderMarketConditionId}:${orderOutcomeIndex}`
-  }, [orderEventId, orderMarketConditionId, orderOutcomeIndex])
-  const fallbackButtonFromOrderState = useMemo(() => {
-    if (orderEventId !== activeCard.event.id || !orderMarketConditionId) {
-      return null
-    }
-
-    if (orderOutcomeIndex === OUTCOME_INDEX.YES || orderOutcomeIndex === OUTCOME_INDEX.NO) {
-      const exactButton = activeCard.buttons.find(button =>
-        button.conditionId === orderMarketConditionId && button.outcomeIndex === orderOutcomeIndex,
-      )
-      if (exactButton) {
-        return exactButton.key
-      }
-    }
-
-    const conditionButton = activeCard.buttons.find(button => button.conditionId === orderMarketConditionId)
-    return conditionButton?.key ?? null
-  }, [activeCard.buttons, activeCard.event.id, orderEventId, orderMarketConditionId, orderOutcomeIndex])
-
-  const activeTradeContext = useMemo(() => {
-    const candidateKeys = usesSectionLayout
-      ? [
-          activeTradeButtonKey,
-          openSectionKey ? selectedButtonBySection[openSectionKey] : null,
-          openAuxiliaryConditionId ? selectedAuxiliaryButtonByConditionId[openAuxiliaryConditionId] : null,
-          marketSlugToButtonKey,
-          fallbackButtonFromOrderState,
-          renderedAuxiliaryMarketCards[0]?.buttons[0]?.key ?? null,
-          moneylineButtonKey,
-          selectedButtonBySection.spread,
-          selectedButtonBySection.total,
-          selectedButtonBySection.btts,
-          resolveDefaultConditionId(activeCard),
-        ]
-      : [
-          activeTradeButtonKey,
-          openAuxiliaryConditionId ? selectedAuxiliaryButtonByConditionId[openAuxiliaryConditionId] : null,
-          marketSlugToButtonKey,
-          fallbackButtonFromOrderState,
-          renderedAuxiliaryMarketCards[0]?.buttons[0]?.key ?? null,
-          resolveDefaultConditionId(activeCard),
-        ]
-    const effectiveButtonKey = candidateKeys.find((buttonKey) => {
-      if (!buttonKey) {
-        return false
-      }
-
-      return activeCard.buttons.some(button => button.key === buttonKey)
-    }) ?? null
-    if (!effectiveButtonKey) {
-      return null
-    }
-
-    const button = resolveSelectedButton(activeCard, effectiveButtonKey)
-    if (!button) {
-      return null
-    }
-
-    const market = resolveSelectedMarket(activeCard, button.key)
-    if (!market) {
-      return null
-    }
-
-    const outcome = resolveSelectedOutcome(market, button)
-    if (!outcome) {
-      return null
-    }
-
-    return { button, market, outcome }
-  }, [
-    activeTradeButtonKey,
-    activeCard,
-    fallbackButtonFromOrderState,
-    usesSectionLayout,
-    moneylineButtonKey,
-    marketSlugToButtonKey,
-    openAuxiliaryConditionId,
-    openSectionKey,
-    renderedAuxiliaryMarketCards,
-    selectedAuxiliaryButtonByConditionId,
-    selectedButtonBySection,
-  ])
-
-  const activeTradeHeaderContext = useMemo(() => {
-    if (!activeTradeContext) {
-      return null
-    }
-
-    if (!orderMarketConditionId || orderMarketConditionId !== activeTradeContext.market.condition_id) {
-      return activeTradeContext
-    }
-
-    if (orderOutcomeIndex == null) {
-      return activeTradeContext
-    }
-
-    const matchedOutcome = activeTradeContext.market.outcomes.find(
-      outcome => outcome.outcome_index === orderOutcomeIndex,
-    ) ?? activeTradeContext.outcome
-
-    const matchedButton = activeCard.buttons.find(
-      button => (
-        button.conditionId === activeTradeContext.market.condition_id
-        && button.outcomeIndex === orderOutcomeIndex
-      ),
-    ) ?? activeTradeContext.button
-
-    return {
-      ...activeTradeContext,
-      button: matchedButton,
-      outcome: matchedOutcome,
-    }
-  }, [activeTradeContext, activeCard.buttons, orderMarketConditionId, orderOutcomeIndex])
-  const orderPanelOutcomeLabelOverrides = useMemo(
-    () => activeTradeContext
-      ? resolveOrderPanelOutcomeLabelOverrides(
-          activeCard,
-          activeTradeHeaderContext?.market ?? activeTradeContext.market,
-        )
-      : {},
-    [activeCard, activeTradeContext, activeTradeHeaderContext],
-  )
-  const orderPanelOutcomeAccentOverrides = useMemo(
-    () => activeTradeContext
-      ? resolveOrderPanelOutcomeAccentOverrides(
-          activeCard,
-          activeTradeHeaderContext?.market ?? activeTradeContext.market,
-        )
-      : {},
-    [activeCard, activeTradeContext, activeTradeHeaderContext],
-  )
-  const activeTradePrimaryOutcomeIndex = useMemo(() => {
-    if (!activeTradeContext || activeTradeContext.button.marketType !== 'spread') {
-      return null
-    }
-
-    return resolveStableSpreadPrimaryOutcomeIndex(activeCard, activeTradeContext.button.conditionId)
-  }, [activeCard, activeTradeContext])
-
-  return {
-    orderSelectionSyncKey,
-    fallbackButtonFromOrderState,
-    activeTradeContext,
-    activeTradeHeaderContext,
-    orderPanelOutcomeLabelOverrides,
-    orderPanelOutcomeAccentOverrides,
-    activeTradePrimaryOutcomeIndex,
-  }
-}
-
-function useSeriesSpreadPickerSync({
-  selectedSpreadSectionButton,
-  detailMarketByConditionId,
-  setActiveSeriesSpreadPickerNumber,
-}: {
-  selectedSpreadSectionButton: SportsGamesButton | null
-  detailMarketByConditionId: Map<string, SportsGamesCard['detailMarkets'][number]>
-  setActiveSeriesSpreadPickerNumber: (setter: (current: number | null) => number | null) => void
-}) {
-  useEffect(function syncActiveSeriesSpreadPickerToSelection() {
-    const selectedMarket = selectedSpreadSectionButton
-      ? detailMarketByConditionId.get(selectedSpreadSectionButton.conditionId) ?? null
-      : null
-    const selectedNumber = parseEsportsSegmentNumber(selectedMarket)
-    if (selectedNumber == null) {
-      return
-    }
-
-    setActiveSeriesSpreadPickerNumber(current => current === selectedNumber ? current : selectedNumber)
-
-    return function noopSyncActiveSeriesSpreadPickerToSelectionCleanup() {}
-  }, [detailMarketByConditionId, selectedSpreadSectionButton, setActiveSeriesSpreadPickerNumber])
-}
-
-function useEsportsSegmentTabKeySync({
-  hasEsportsSegmentedLayout,
-  marketSlugToButtonKey,
-  activeCard,
-  detailMarketByConditionId,
-  setActiveEsportsSegmentTabKey,
-}: {
-  hasEsportsSegmentedLayout: boolean
-  marketSlugToButtonKey: string | null
-  activeCard: SportsGamesCard
-  detailMarketByConditionId: Map<string, SportsGamesCard['detailMarkets'][number]>
-  setActiveEsportsSegmentTabKey: (value: EsportsLayoutTabKey) => void
-}) {
-  useEffect(function syncEsportsSegmentTabKeyFromMarketSlug() {
-    if (!hasEsportsSegmentedLayout || !marketSlugToButtonKey) {
-      return
-    }
-
-    const selectedButton = activeCard.buttons.find(button => button.key === marketSlugToButtonKey) ?? null
-    const selectedMarket = selectedButton
-      ? detailMarketByConditionId.get(selectedButton.conditionId) ?? null
-      : null
-
-    if (!selectedMarket) {
-      return
-    }
-
-    const mapNumber = parseEsportsSegmentNumber(selectedMarket)
-    if (mapNumber != null) {
-      setActiveEsportsSegmentTabKey(resolveEsportsSegmentTabKey(mapNumber))
-      return
-    }
-
-    setActiveEsportsSegmentTabKey('series')
-
-    return function noopSyncEsportsSegmentTabKeyFromMarketSlugCleanup() {}
-  }, [activeCard.buttons, detailMarketByConditionId, hasEsportsSegmentedLayout, marketSlugToButtonKey, setActiveEsportsSegmentTabKey])
-}
-
-function useOrderStateSync({
-  activeCard,
-  activeTradeContextButtonKey,
-  orderEventId,
-  orderMarketConditionId,
-  orderOutcomeIndex,
-  marketSlugToButtonKey,
-  fallbackButtonFromOrderState,
-  orderSelectionSyncKey,
-  usesSectionLayout,
-  detailMarketByConditionId,
-  auxiliaryPanelKeyByButtonKey,
-  setActiveTradeButtonKey,
-  setSelectedButtonBySection,
-  setSelectedAuxiliaryButtonByConditionId,
-  setOrderEvent,
-  setOrderMarket,
-  setOrderOutcome,
-  setOrderSide,
-}: {
-  activeCard: SportsGamesCard
-  activeTradeContextButtonKey: string | null
-  orderEventId: string | null
-  orderMarketConditionId: string | null
-  orderOutcomeIndex: number | null
-  marketSlugToButtonKey: string | null
-  fallbackButtonFromOrderState: string | null
-  orderSelectionSyncKey: string | null
-  usesSectionLayout: boolean
-  detailMarketByConditionId: Map<string, SportsGamesCard['detailMarkets'][number]>
-  auxiliaryPanelKeyByButtonKey: Map<string, string>
-  setActiveTradeButtonKey: React.Dispatch<React.SetStateAction<string | null>>
-  setSelectedButtonBySection: React.Dispatch<React.SetStateAction<Record<EventSectionKey, string | null>>>
-  setSelectedAuxiliaryButtonByConditionId: React.Dispatch<React.SetStateAction<Record<string, string | null>>>
-  setOrderEvent: (event: SportsGamesCard['event']) => void
-  setOrderMarket: (market: SportsGamesCard['detailMarkets'][number]) => void
-  setOrderOutcome: (outcome: SportsGamesCard['detailMarkets'][number]['outcomes'][number]) => void
-  setOrderSide: (side: typeof ORDER_SIDE[keyof typeof ORDER_SIDE]) => void
-}) {
-  const pushedOrderSelectionRef = useRef<string | null>(null)
-  const currentOrderSelectionRef = useRef<{
-    eventId: string | null
-    conditionId: string | null
-    outcomeIndex: number | null
-  }>({
-    eventId: null,
-    conditionId: null,
-    outcomeIndex: null,
-  })
-
-  useEffect(function applyFallbackOrderStateToSelection() {
-    if (marketSlugToButtonKey) {
-      return
-    }
-
-    if (orderSelectionSyncKey && orderSelectionSyncKey === pushedOrderSelectionRef.current) {
-      return
-    }
-
-    if (!fallbackButtonFromOrderState) {
-      return
-    }
-
-    const matchedButton = activeCard.buttons.find(
-      button => button.key === fallbackButtonFromOrderState,
-    )
-    if (!matchedButton) {
-      return
-    }
-
-    setActiveTradeButtonKey((current) => {
-      if (current === matchedButton.key) {
-        return current
-      }
-      return matchedButton.key
-    })
-
-    if (usesSectionLayout) {
-      const matchedMarket = detailMarketByConditionId.get(matchedButton.conditionId) ?? null
-      const sectionKey = resolveEventSectionKeyForButton(matchedButton, matchedMarket)
-      if (sectionKey) {
-        setSelectedButtonBySection((current) => {
-          if (current[sectionKey] === matchedButton.key) {
-            return current
-          }
-
-          return {
-            ...current,
-            [sectionKey]: matchedButton.key,
-          }
-        })
-        return
-      }
-    }
-
-    setSelectedAuxiliaryButtonByConditionId((current) => {
-      const auxiliaryPanelKey = auxiliaryPanelKeyByButtonKey.get(matchedButton.key) ?? matchedButton.conditionId
-
-      if (current[auxiliaryPanelKey] === matchedButton.key) {
-        return current
-      }
-
-      return {
-        ...current,
-        [auxiliaryPanelKey]: matchedButton.key,
-      }
-    })
-
-    return function noopApplyFallbackOrderStateToSelectionCleanup() {}
-  }, [
-    activeCard.buttons,
-    auxiliaryPanelKeyByButtonKey,
-    detailMarketByConditionId,
-    fallbackButtonFromOrderState,
-    marketSlugToButtonKey,
-    orderSelectionSyncKey,
-    usesSectionLayout,
-    setActiveTradeButtonKey,
-    setSelectedAuxiliaryButtonByConditionId,
-    setSelectedButtonBySection,
-  ])
-
-  useEffect(function trackCurrentOrderSelection() {
-    currentOrderSelectionRef.current = {
-      eventId: orderEventId,
-      conditionId: orderMarketConditionId,
-      outcomeIndex: orderOutcomeIndex,
-    }
-  }, [orderEventId, orderMarketConditionId, orderOutcomeIndex])
-
-  useEffect(function pushActiveTradeSelectionToOrderState() {
-    if (!activeTradeContextButtonKey) {
-      pushedOrderSelectionRef.current = null
-      return
-    }
-
-    const button = resolveSelectedButton(activeCard, activeTradeContextButtonKey)
-    const market = resolveSelectedMarket(activeCard, activeTradeContextButtonKey)
-    const outcome = resolveSelectedOutcome(market, button)
-    if (!button || !market || !outcome) {
-      pushedOrderSelectionRef.current = null
-      return
-    }
-
-    const nextOrderSelectionSyncKey = `${activeCard.event.id}:${market.condition_id}:${outcome.outcome_index}`
-    const {
-      eventId: currentOrderEventId,
-      conditionId: currentOrderMarketConditionId,
-      outcomeIndex: currentOrderOutcomeIndex,
-    } = currentOrderSelectionRef.current
-
-    if (
-      currentOrderEventId === activeCard.event.id
-      && currentOrderMarketConditionId === market.condition_id
-      && currentOrderOutcomeIndex === outcome.outcome_index
-    ) {
-      pushedOrderSelectionRef.current = nextOrderSelectionSyncKey
-      return
-    }
-
-    pushedOrderSelectionRef.current = nextOrderSelectionSyncKey
-    setOrderEvent(activeCard.event)
-    setOrderMarket(market)
-    setOrderOutcome(outcome)
-    setOrderSide(ORDER_SIDE.BUY)
-
-    return function noopPushActiveTradeSelectionToOrderStateCleanup() {}
-  }, [
-    activeCard,
-    activeTradeContextButtonKey,
-    setOrderEvent,
-    setOrderMarket,
-    setOrderOutcome,
-    setOrderSide,
-  ])
-}
-
-function useSeriesSegmentPickerData({
-  activeCard,
-  groupedButtons,
-  detailMarketByConditionId,
-  hasEsportsSegmentedLayout,
-  activeSeriesSpreadPickerNumber,
-  selectedSpreadSectionButton,
-  selectedTotalSectionButton,
-}: {
-  activeCard: SportsGamesCard
-  groupedButtons: Record<SportsGamesMarketType, SportsGamesButton[]>
-  detailMarketByConditionId: Map<string, SportsGamesCard['detailMarkets'][number]>
-  hasEsportsSegmentedLayout: boolean
-  activeSeriesSpreadPickerNumber: number | null
-  selectedSpreadSectionButton: SportsGamesButton | null
-  selectedTotalSectionButton: SportsGamesButton | null
-}) {
-  const seriesSpreadSegmentOptions = useMemo(() => {
-    if (!hasEsportsSegmentedLayout) {
-      return [] as Array<{
-        number: number
-        conditionIds: string[]
-        buttonsByConditionId: Map<string, SportsGamesButton[]>
-      }>
-    }
-
-    const byNumber = new Map<number, {
-      number: number
-      conditionIds: string[]
-      buttonsByConditionId: Map<string, SportsGamesButton[]>
-    }>()
-
-    groupedButtons.spread.forEach((button) => {
-      const market = detailMarketByConditionId.get(button.conditionId) ?? null
-      const number = parseEsportsSegmentNumber(market)
-      if (number == null) {
-        return
-      }
-
-      const existing = byNumber.get(number)
-      if (existing) {
-        if (!existing.conditionIds.includes(button.conditionId)) {
-          existing.conditionIds.push(button.conditionId)
-        }
-        const currentButtons = existing.buttonsByConditionId.get(button.conditionId) ?? []
-        currentButtons.push(button)
-        existing.buttonsByConditionId.set(button.conditionId, currentButtons)
-        return
-      }
-
-      byNumber.set(number, {
-        number,
-        conditionIds: [button.conditionId],
-        buttonsByConditionId: new Map([[button.conditionId, [button]]]),
-      })
-    })
-
-    return Array.from(byNumber.values())
-      .map(option => ({
-        ...option,
-        buttonsByConditionId: new Map(
-          Array.from(option.buttonsByConditionId.entries())
-            .map(([conditionId, buttons]) => [conditionId, sortSectionButtons('spread', buttons)] as const),
-        ),
-      }))
-      .sort((left, right) => left.number - right.number)
-  }, [detailMarketByConditionId, groupedButtons.spread, hasEsportsSegmentedLayout])
-  const activeSeriesSpreadSegmentOption = useMemo(() => {
-    if (seriesSpreadSegmentOptions.length === 0) {
-      return null
-    }
-
-    if (activeSeriesSpreadPickerNumber != null) {
-      const byPickerNumber = seriesSpreadSegmentOptions.find(option => option.number === activeSeriesSpreadPickerNumber)
-      if (byPickerNumber) {
-        return byPickerNumber
-      }
-    }
-
-    const selectedMarket = selectedSpreadSectionButton
-      ? detailMarketByConditionId.get(selectedSpreadSectionButton.conditionId) ?? null
-      : null
-    const selectedNumber = parseEsportsSegmentNumber(selectedMarket)
-    if (selectedNumber != null) {
-      return seriesSpreadSegmentOptions.find(option => option.number === selectedNumber)
-        ?? seriesSpreadSegmentOptions[0]
-        ?? null
-    }
-
-    return seriesSpreadSegmentOptions[0] ?? null
-  }, [
-    activeSeriesSpreadPickerNumber,
-    detailMarketByConditionId,
-    selectedSpreadSectionButton,
-    seriesSpreadSegmentOptions,
-  ])
-  const activeSeriesSpreadSegmentNumber = activeSeriesSpreadSegmentOption?.number ?? null
-  const activeSeriesSpreadConditionId = useMemo(() => {
-    if (!activeSeriesSpreadSegmentOption) {
-      return null
-    }
-
-    const currentSpreadConditionId = selectedSpreadSectionButton?.conditionId ?? null
-    if (
-      currentSpreadConditionId
-      && activeSeriesSpreadSegmentOption.buttonsByConditionId.has(currentSpreadConditionId)
-    ) {
-      return currentSpreadConditionId
-    }
-
-    return activeSeriesSpreadSegmentOption.conditionIds[0] ?? null
-  }, [activeSeriesSpreadSegmentOption, selectedSpreadSectionButton])
-  const seriesSpreadSegmentPickerOptions = useMemo(
-    () => seriesSpreadSegmentOptions.map(option => ({
-      key: `spread-segment-${option.number}`,
-      label: `${option.number}`,
-      number: option.number,
-    })),
-    [seriesSpreadSegmentOptions],
-  )
-  const seriesTotalLinePickerOptions = useMemo(() => {
-    if (!hasEsportsSegmentedLayout) {
-      return [] as SportsLinePickerOption[]
-    }
-
-    const allowedConditionIds = new Set(groupedButtons.total.map(button => button.conditionId))
-    return buildLinePickerOptions(activeCard, 'total')
-      .filter(option => allowedConditionIds.has(option.conditionId))
-  }, [activeCard, groupedButtons.total, hasEsportsSegmentedLayout])
-  const activeSeriesTotalLineOption = useMemo(() => {
-    if (seriesTotalLinePickerOptions.length === 0) {
-      return null
-    }
-
-    if (selectedTotalSectionButton) {
-      return seriesTotalLinePickerOptions.find(option => option.conditionId === selectedTotalSectionButton.conditionId)
-        ?? seriesTotalLinePickerOptions[0]
-        ?? null
-    }
-
-    return seriesTotalLinePickerOptions[0] ?? null
-  }, [selectedTotalSectionButton, seriesTotalLinePickerOptions])
-  const activeSeriesTotalConditionId = activeSeriesTotalLineOption?.conditionId ?? null
-  const activeSeriesTotalLineValue = activeSeriesTotalLineOption?.lineValue ?? null
-  const seriesTotalPickerOptions = useMemo(
-    () => seriesTotalLinePickerOptions.map(option => ({
-      key: `total-line-${option.conditionId}`,
-      label: option.label,
-      number: option.lineValue,
-    })),
-    [seriesTotalLinePickerOptions],
-  )
-
-  return {
-    seriesSpreadSegmentOptions,
-    activeSeriesSpreadSegmentOption,
-    activeSeriesSpreadSegmentNumber,
-    activeSeriesSpreadConditionId,
-    seriesSpreadSegmentPickerOptions,
-    seriesTotalLinePickerOptions,
-    activeSeriesTotalLineOption,
-    activeSeriesTotalConditionId,
-    activeSeriesTotalLineValue,
-    seriesTotalPickerOptions,
-  }
-}
 
 export default function SportsEventCenter({
   card,
@@ -3929,7 +483,7 @@ export default function SportsEventCenter({
   const esportsSegmentTabs = hasEsportsSegmentedLayout
     ? [
         { key: 'series' as const, label: 'Series Lines' },
-        ...esportsSegmentTabNumbers.map(mapNumber => ({ key: resolveEsportsSegmentTabKey(mapNumber), label: `${segmentLabel} ${mapNumber}` })),
+        ...esportsSegmentTabNumbers.map(mapNumber => ({ key: `segment-${mapNumber}` as const, label: `${segmentLabel} ${mapNumber}` })),
       ]
     : []
   const marketViewTabs = hasMultipleMarketViews
